% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitokens.r
\name{add_multitoken_label}
\alias{add_multitoken_label}
\title{Choose and add multitoken strings based on multitoken categories}
\usage{
add_multitoken_label(
  tc,
  colloc_id,
  feature = "token",
  new_feature = sprintf("\%s_l", colloc_id),
  pref_subset = NULL
)
}
\arguments{
\item{tc}{a tcorpus object}

\item{colloc_id}{the data column containing the unique id for multitoken tokens}

\item{feature}{the name of the feature column}

\item{new_feature}{the name of the new feature column}

\item{pref_subset}{Optionally, a subset call, to specify a subset that has priority for finding the most frequently occuring string}
}
\description{
Given a multitoken category (e.g., named entity ids), this function finds the most frequently occuring string in this category and adds it as a label for the category
}
