% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcorpus.r
\name{as.tcorpus.default}
\alias{as.tcorpus.default}
\title{Force an object to be a tCorpus class}
\usage{
\method{as.tcorpus}{default}(x, ...)
}
\arguments{
\item{x}{the object to be forced}

\item{...}{not used}
}
\description{
Force an object to be a tCorpus class
}

