% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_tcorpus.r
\name{as.tcorpus}
\alias{as.tcorpus}
\title{Force an object to be a tCorpus class}
\usage{
as.tcorpus(x, ...)
}
\arguments{
\item{x}{the object to be forced}

\item{...}{not used}
}
\description{
Force an object to be a tCorpus class
}
