% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_tcorpus.r
\name{as.tcorpus.tCorpus}
\alias{as.tcorpus.tCorpus}
\title{Force an object to be a tCorpus class}
\usage{
\method{as.tcorpus}{tCorpus}(x, ...)
}
\arguments{
\item{x}{the object to be forced}

\item{...}{not used}
}
\description{
Force an object to be a tCorpus class
}
\examples{
tc = create_tcorpus(c('First text', 'Second text'))
as.tcorpus(tc)
}
