% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semnet_backbone.r
\name{backbone_alpha}
\alias{backbone_alpha}
\title{Calculate the alpha values that can be used to extract the backbone of a network.}
\usage{
backbone_alpha(g, k.is.Nvertices = F)
}
\arguments{
\item{g}{A graph in the `Igraph` format.}
}
\value{
A vector of alpha values, which matches the edges. Can thus easily be made an edge attribute: E(g)$alpha = backbone.alpha(g)
}
\description{
Based on the following paper: Serrano, M. Á., Boguñá, M., & Vespignani, A. (2009). Extracting the multiscale backbone of complex weighted networks. Proceedings of the National Academy of Sciences, 106(16), 6483-6488.
}

