% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_functions.r
\name{calc_chi2}
\alias{calc_chi2}
\title{Compute the chi^2 statistic for a 2x2 crosstab containing the values
[[a, b], [c, d]]}
\usage{
calc_chi2(a, b, c, d, autocorrect = T, yates_correction = rep(F, length(a)))
}
\description{
Compute the chi^2 statistic for a 2x2 crosstab containing the values
[[a, b], [c, d]]
}

