% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicbrowser.r
\name{clusterinfo}
\alias{clusterinfo}
\title{Create a 'clusterinfo' object that can be used for the individual render methods}
\usage{
clusterinfo(m, terms, documents, meta, topic_ids = 1:m@k, words = terms,
  date_interval = "year")
}
\arguments{
\item{terms:}{a vector of terms, which should be sorted in their original order and match m@terms}

\item{documents:}{a vector of the same length as terms, indicating the document of each term, matching m@documents}

\item{meta:}{a data frame with meta data about the documents, should have columns aid, headline, medium, and date (todo: make more flexible)}

\item{topic_ids:}{optionally restrict output to a selection of topics}

\item{date_interval:}{specify the interval for plotting the meta$data}

\item{words:}{if given, use instead of terms for displaying document}
}
\value{
a list with tokens, wordassignments, and other items
}
\description{
#' @param m: the fitted LDA model object from the topicmodels package
}

