% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{code_features}
\alias{code_features}
\title{Annotate a data frame of tokens with codes using Lucene-like search queries}
\usage{
code_features(tc, queries, batchsize = 5, default_window = NA,
  condition_once = FALSE, keyword_filter = rep(T, nrow(tokens)),
  presorted = F, doc.col = getOption("doc.col", "doc_id"),
  position.col = getOption("position.col", "position"),
  feature.col = getOption("feature.col", "word"), verbose = T)
}
\arguments{
\item{queries}{a data frame containing the queries.}

\item{batchsize}{This function is faster if multiple queries are searched together, but too many queries (with too many tokens) at once can eat up memory or crash R. Try lowering batchsize in case of issues.}

\item{default_window}{Determines the default word distance of the condition terms to the keyword (thus, if no specific word distance is set with the ~ symbol)}

\item{condition_once}{logical. If TRUE, then if an keyword satisfies its conditions once in an article, all keywords within that article are coded.}

\item{keyword_filter}{A logical vector that indicates which tokens can match an keyword. Can for instance be used to only select tokens that are proper names (using POS tagging) when looking for people.}

\item{presorted}{The data has to be sorted on order(doc_id, position). If this is already the case, presorted can be set to TRUE to save time (which is usefull when testing many individual queries for large tokenlists)}

\item{doc.col}{The name of the document_id column. Defaults to "doc_id", unless a global default is specified using setTokenlistColnames()}

\item{position.col}{The name of the column giving the position in a document. Defaults to "position", unless a global default is specified using setTokenlistColnames()}

\item{feature.col}{The name of the column containing the token feature. Defaults to "word", unless a global default is specified using setTokenlistColnames()}

\item{verbose}{show progress}

\item{tokens}{a tokenlist object. See ?asTokenlist() for details.}
}
\value{
the annotated tokens data frame
}
\description{
Annotate a data frame of tokens with codes using Lucene-like search queries
}

