% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semnet.r
\name{cooccurrence}
\alias{cooccurrence}
\title{Create a semantic network based on the co-occurence of words in documents}
\usage{
cooccurrence(tc, feature, measure = c("cooccurrence", "cosine", "conprob",
  "conprob_weighted"), context_level = c("document", "sentence"))
}
\arguments{
\item{tc}{A tCorpus object}

\item{measure}{the measure to calculate adjacency. Options are "cooccurrence" (number of documents), "cosine" (cosine similarity), "conprob" (conditional probability) or "conprob_weighted". Of these "cooccurrence" and "cosine" are symetric (word1 -> word2 is equal to word2 -> word1) whereas "conprob" is assymetric. "conprob_weighted" is an extension of "conprob" that takes into account how often words occur in documents, by first transforming the number of times the word occurs into a probability.}

\item{feature.col}{Optional, a character string indicating the column of the tokenlist data that is used to represent the token feature (e.g., words, lemma, ngrams). If NULL, the feature.col slot of the tokenlist object will be used.}
}
\value{
A graph in the Igraph format in which edges represent the adjacency of terms
}
\description{
This function calculates the co-occurence of words and returns a network/graph where nodes are words and edges represent the similarity/adjacency of words. Co-occurence is calcuated based on how often two words occured within the same document (e.g., news article, chapter, paragraph, sentence). Note that the cooc_window() function can be used to calculate co-occurrence of words within a given word distance.
}

