% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_functions.r
\name{cooccurrence_matrix_window}
\alias{cooccurrence_matrix_window}
\title{Calculate the cooccurrence of features.}
\usage{
cooccurrence_matrix_window(tc, feature, matrix_mode = "position_to_window",
  count_mode = "dicho", mat_stats = c("sum.x", "sum.y", "magnitude.x",
  "magnitude.y"), context_level = "document", direction = "<>",
  window.size = 10, n.batches = window.size, alpha = 2,
  drop_empty_terms = T)
}
\arguments{
\item{tc}{a tcorpus object}

\item{mat_stats}{A character vector indicating which matrix statistics to calculate. These are necessary for calculating similarity/distance metrics.}

\item{alpha}{}

\item{mode}{Either 'window_to_window' or 'position_to_window'. With 'window_to_window', cooccurrence is calculated as the number of times the window of two features overlaps. With 'position_to_window', cooccurrence is calculated as the number of times a feature occurs (is positioned) within the window of another feature}
}
\value{
a list with the cooccurrence matrix and the specified matrix statistics (mat_stats)
}
\description{
Returns a list with the cooccurrence matrix, and the matrix statistics specified in mat_stats.
}
\details{
The reason for making this a separate, complex function is that it enables the construction of the window matrix and the matrix multiplication to be executed in batches/shards
}

