% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semnet.r
\name{cooccurrence_window}
\alias{cooccurrence_window}
\title{A sliding window approach to calculate the co-occurence of words}
\usage{
cooccurrence_window(tc, feature, measure = c("con_prob", "cosine",
  "count_directed", "count_undirected"), context_level = c("document",
  "sentence"), window.size = 10, direction = "<>", chi2 = F,
  backbone = F)
}
\arguments{
\item{backbone}{}
}
\description{
A sliding window approach to calculate the co-occurence of words
}

