% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicbrowser.r
\name{createTopicBrowser}
\alias{createTopicBrowser}
\title{Create a topic browser}
\usage{
createTopicBrowser(x, terms = NULL, documents = NULL, meta = NULL,
  topic_ids = 1:m@k, date_interval = "year", words = terms,
  output = NULL, browse = interactive())
}
\arguments{
\item{m:}{the fitted LDA model object from the topicmodels package}

\item{terms:}{a vector of terms, which should be sorted in their original order and match m@terms}

\item{documents:}{a vector of the same length as terms, indicating the document of each term, matching m@documents}

\item{meta:}{a data frame with meta data about the documents, should have columns aid, headline, medium, and date (todo: make more flexible)}

\item{topic_ids:}{optionally restrict output to a selection of topics}

\item{date_interval:}{specify the interval for plotting the meta$data}

\item{words:}{if given, use instead of terms for displaying document}

\item{output:}{an optional file name to write the browser html to}
}
\value{
the html generated or (invisible) the filename
}
\description{
This will create a set of linked html pages to browse the given topic model
}

