% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.r, R/resources.r
\name{download_resource}
\alias{download_resource}
\title{Download lookup lists from various resources}
\usage{
download_resource(resource = c("jrc_names"), save_path = tempdir(),
  force_download = F)

download_resource(resource = c("jrc_names"), save_path = tempdir(),
  force_download = F)
}
\arguments{
\item{resource}{A character string indicating which resource to download}

\item{save_path}{The path (without the filename) where the resource is stored. Default is a temporary directory, that is cleaned up when the R session ends.}

\item{force_download}{If TRUE, the resource will be downloaded even if it is stored locally.}

\item{resource}{A character string indicating which resource to download}

\item{save_path}{The path (without the filename) where the resource is stored. Default is a temporary directory, that is cleaned up when the R session ends.}

\item{force_download}{If TRUE, the resource will be downloaded even if it is stored locally.}
}
\description{
Currently available resources:
\itemize{
   \item{JRC-Names: "JRC-Names is a highly multilingual named entity resource for person and organisation names. [...] JRC-Names is a by-product of the analysis of about 220,000 news reports per day by the Europe Media Monitor (EMM) family of applications." (https://ec.europa.eu/jrc/en/language-technologies/jrc-names)}
}

Currently available resources:
\itemize{
   \item{JRC-Names: "JRC-Names is a highly multilingual named entity resource for person and organisation names. [...] JRC-Names is a by-product of the analysis of about 220,000 news reports per day by the Europe Media Monitor (EMM) family of applications." (https://ec.europa.eu/jrc/en/language-technologies/jrc-names)}
}
}
\details{
The resource file is downloaded and saved as a temporary file, which will be used if this function is again called within the same R session.
Alternatively, the file can be stored locally by specifying the path. This is a kind thing to do if you mean to use this resource regularly (and do not need it to be up to date).

The resource file is downloaded and saved as a temporary file, which will be used if this function is again called within the same R session.
Alternatively, the file can be stored locally by specifying the path. This is a kind thing to do if you mean to use this resource regularly (and do not need it to be up to date).
}

