% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_information.r
\name{dtm_compare}
\alias{dtm_compare}
\title{Compare two corpora}
\usage{
dtm_compare(dtm.x, dtm.y = NULL, smooth = 1, min.over = NULL,
  min.chi = NULL, select.rows = NULL, x_is_subset = F)
}
\arguments{
\item{dtm.x}{the main document-term matrix}

\item{dtm.y}{the 'reference' document-term matrix}

\item{smooth}{the smoothing parameter for Laplace smoothing. for computing overrepresentation. Currently using Laplace smoothing.}
}
\value{
A data frame with rows corresponding to the terms in dtm and the statistics in the columns
}
\description{
Compare the term use in corpus dtm with a refernece corpus dtm.ref, returning relative frequencies
and overrepresentation using various measures.
}
\details{
Note that the relative frequency (relfreq) and over-ratio (over) are based on smoothed scores, to prevent the over ratio to be 0 or Inf. Smoothing can be turned of by setting smooth to 0
}

