% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{emoticon_dict}
\alias{emoticon_dict}
\title{Dictionary with common ASCII emoticons}
\format{A data.frame with a "string" and "code" column.}
\usage{
data(emoticon_dict)
}
\description{
Obtained from the Wikipedia List_of_emoticons page.
}
\keyword{datasets}
