% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_lists.r, R/resources.r
\name{feature_from_resource}
\alias{feature_from_resource}
\title{Use a language resource to create a feature}
\usage{
feature_from_resource(tc, resource = "jrc_names", new_feature = resource,
  feature = "word", save_path = tempdir(), force_download = F,
  multiword_labels = T, verbose = T)

feature_from_resource(tc, resource = "jrc_names", new_feature = resource,
  feature = "word", save_path = tempdir(), force_download = F,
  multiword_labels = T, verbose = T)
}
\arguments{
\item{tc}{a tCorpus object}

\item{resource}{A character string indicating which resource to use.}

\item{new_feature}{The name of the new feature. Default is the name of the resource}

\item{feature}{The feature to be used as input. Note that most resources assume that the input consists of unprocessed words}

\item{save_path}{The path (without the filename) where the resource is stored. Default is a temporary directory, that is cleaned up when the R session ends.}

\item{force_download}{If TRUE, the resource will be downloaded even if it is stored locally.}

\item{multiword_labels}{if True, then for resources that create an id for subsequent words (e.g., named entities), a label column is added based on the most frequently occuring multiword combination in the data.}

\item{tc}{a tCorpus object}

\item{resource}{A character string indicating which resource to use.}

\item{feature}{The feature to be used as input. Note that most resources assume that the input consists of unprocessed words}

\item{save_path}{The path (without the filename) where the resource is stored. Default is a temporary directory, that is cleaned up when the R session ends.}

\item{force_download}{If TRUE, the resource will be downloaded even if it is stored locally.}
}
\description{
Currently available resources:
\itemize{
   \item{JRC-Names: "JRC-Names is a highly multilingual named entity resource for person and organisation names. [...] JRC-Names is a by-product of the analysis of about 220,000 news reports per day by the Europe Media Monitor (EMM) family of applications." (https://ec.europa.eu/jrc/en/language-technologies/jrc-names)}
}

Currently available resources:
\itemize{
   \item{JRC-Names: "JRC-Names is a highly multilingual named entity resource for person and organisation names. [...] JRC-Names is a by-product of the analysis of about 220,000 news reports per day by the Europe Media Monitor (EMM) family of applications." (https://ec.europa.eu/jrc/en/language-technologies/jrc-names)}
}
}
\details{
ABOUT THE RESOURCE DATA:
The resource file is downloaded and saved as a temporary file, which will be used if this function is again called within the same R session.
Alternatively, the file can be stored locally by specifying the path. This is a kind thing to do if you mean to use this resource regularly (and do not need it to be up to date).

ABOUT THE RESOURCE DATA:
The resource file is downloaded and saved as a temporary file, which will be used if this function is again called within the same R session.
Alternatively, the file can be stored locally by specifying the path. This is a kind thing to do if you mean to use this resource regularly (and do not need it to be up to date).
}

