% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.r
\name{fill_time_gaps}
\alias{fill_time_gaps}
\title{Add empty values for pretty plotting}
\usage{
fill_time_gaps(d, date_interval)
}
\arguments{
\item{d}{A data.frame with the columns 'time' (Date) and 'value' (numeric)}

\item{date_interval}{The date_interval is required to know what the gaps are}
}
\value{
A data.frame with the columns 'time' (Date) and 'value' (numeric)
}
\description{
When plotting a timeline, gaps in date_intervals are ignored. For the attention for topics gaps should be considered as having value 0.
}

