% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcorpus.r
\name{filter_tcorpus}
\alias{filter_tcorpus}
\title{Title}
\usage{
filter_tcorpus(tc, filter_keep = NULL, filter_ignore = NULL,
  feature = NULL, language = "english", ignore_stopwords = T,
  ignore_punctuation = T, ignore_features = NULL, keep_features = NULL,
  max_word_i = NULL, max_sent_i = NULL, ignore_doc_id = NULL,
  keep_doc_id = NULL, reset_feature_matrix = T, keep_filtered_data = T)
}
\arguments{
\item{tc}{tcorpus object}

\item{filter_keep}{a custom filter indicating which data to keep. If given, has to be a logical vector or a numerical vector with indices of the same length as the tcorpus data. The custom filter will override}

\item{filter_ignore}{a custom filter indicating which data to ignore. If given, has to be a logical vector or a numerical vector with indices of the same length as the tcorpus data.}

\item{reset_feature_matrix}{}
}
\description{
Title
}

