% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_term_matrix.r
\name{get_dtm}
\alias{get_dtm}
\title{Create a document term matrix from a tCorpus}
\usage{
get_dtm(tc, feature, context_level = c("document", "sentence"),
  weight = c("termfreq", "docfreq", "tfidf", "norm_tfidf"),
  drop_empty_terms = T, form = c("Matrix", "tm_dtm", "quanteda_dfm"),
  subset_tokens = NULL, subset_meta = NULL, context = NULL,
  context_labels = T, feature_labels = T, ngrams = NA,
  ngram_before_subset = F, env = environment())
}
\arguments{
\item{context_level}{}
}
\description{
Create a document term matrix from a tCorpus
}

