% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcorpus.r
\name{get_feature_index}
\alias{get_feature_index}
\title{Get a data.table for quick lookup of features}
\usage{
get_feature_index(tc, feature = "word", context_level = "document",
  max_window_size = 100)
}
\arguments{
\item{max_window_size}{}
}
\description{
#' The feature index has 3 columns:
\itemize{
   \item{feature is a factor of a given feature column, and is set as the key of the data.table}
   \item{i is the index of the rows in tc@data.}
   \item{global_i is an index for features with gaps between contexts (e.g., documents, sentences) to enable word disstance computations using a single vector.}
}
}
\details{
Automatically filters out features based on the filter columns in tc@data
}

