% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_functions.r
\name{get_global_i}
\alias{get_global_i}
\title{Compute global feature positions}
\usage{
get_global_i(tc, context_level = c("document", "sentence"),
  max_window_size = 200)
}
\arguments{
\item{tc}{tCorpus object}

\item{context_level}{The context within which}

\item{max_window_size}{}
}
\value{
a tCorpus object
}
\description{
Features are given global ids, with an added distance (max_window_size) between contexts (e.g., documents, sentences).
This way, the distance of features can be calculated across multiple contexts using a single vector
}

