% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_functions.r
\name{get_query_matrix}
\alias{get_query_matrix}
\title{Creates a matrix where rows match the tokens, columns represent the queries in query_regex, and values represent the word distance to each query hit}
\usage{
get_query_matrix(fdt, query_regex, default.window = NA, return_i = NULL)
}
\arguments{
\item{default.window}{}
}
\value{
a sparse matrix
}
\description{
Creates a matrix where rows match the tokens, columns represent the queries in query_regex, and values represent the word distance to each query hit
}

