% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.r
\name{get_stopwords}
\alias{get_stopwords}
\title{Get a character vector of stopwords}
\usage{
get_stopwords(lang)
}
\arguments{
\item{lang}{The language. Current options are: "danish", "dutch", "english", "finnish", "french", "german", "hungarian", "italian", "norwegian", "portuguese", "romanian", "russian", "spanish" and "swedish"}
}
\value{
A character vector containing stopwords
}
\description{
Get a character vector of stopwords
}

