% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_term_matrix.r
\name{getDTM}
\alias{getDTM}
\title{Create a document term matrix from a tCorpus}
\usage{
getDTM(tc, feature, context_level = c("document", "sentence"))
}
\arguments{
\item{context_level}{}
}
\description{
Create a document term matrix from a tCorpus
}

