% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_results.r
\name{getKwic}
\alias{getKwic}
\title{Get keyword-in-confeature from a token list}
\usage{
getKwic(tc, hits = NULL, token_i = NULL, nwords = 10, nsample = NA,
  prettypaste = T)
}
\arguments{
\item{nwords}{the number of words in front and after the keyword}

\item{tokens}{a data frame of tokens containing columns for document id (doc_id), feature position (position) and feature string (column name can be specified in feature.col, defaults to 'word').}

\item{feature.col}{a character string giving the name of the term string column}
}
\value{
A data.frame with the keyword in confeature
}
\description{
Get keyword-in-confeature from a token list
}

