% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_document.r
\name{html_tagged_text}
\alias{html_tagged_text}
\title{Render text as HTML}
\usage{
html_tagged_text(tc, context_level = "document", text_col = "word",
  meta_columns = NULL, highlight = NULL, scale = NULL, index = NULL,
  highlight_col = "yellow", scale_col = c("red", "blue"))
}
\arguments{
\item{tc}{a tCorpus object}

\item{context_level}{Either 'document' or 'sentence'}

\item{text_col}{The column that is used to create the output text}

\item{meta_columns}{Meta columns are presented in the HTML as a table. If meta_columns is NULL, all meta columns are used. Specific columns can be selected by giving a character vector with meta column names}

\item{highlight}{A numeric or logical vector of length tc$n. Words that are TRUE, not NA and not zero will be highlighted. If highlight is a numeric vector, values determine the alpha of the highlight color (must be between 0 and 1).}

\item{scale}{Like highlight, but uses a color scale. Must be a numeric vector with values ranging from -1 to 1.}

\item{index}{A vector of length tc$n. Unique values will be used to assign words to categories. NA values can be used to leave words unassigned.}

\item{highlight_col}{a character vector of lenght 1 with a color name. Determines the color used in highlight}

\item{scale_col}{a character vector of length 2 with color names, giving the left and right color of the scale}
}
\description{
Render text as HTML
}

