% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc_names.r
\name{jrc_names}
\alias{jrc_names}
\title{Multilingual named entity recognition using the JRC-NAMES resource}
\usage{
jrc_names(tc, new_feature = "jrc_names", feature = "word",
  resource_path = getOption("tcorpus_resources", NULL),
  collocation_labels = T, batchsize = 50000, low_memory = T,
  verbose = T)
}
\arguments{
\item{tc}{a tCorpus object}

\item{new_feature}{The column name of the new feature.}

\item{feature}{The feature to be used as input. For JRC names regular (unprocessed) words should be used.}

\item{resource_path}{The path (without the filename) where the resource is stored. See ?download_resource for more information.}

\item{collocation_labels}{if True, then for resources that create an id for subsequent words (e.g., named entities), labels are added (in a separate column) based on the most frequent collocation combinations in 'your' data. Note that this means that the labels can be different if you run the same analysis on a different corpus; this is why the id is always kept.}

\item{batchsize}{The number of named entity string variations per batch. Using bigger batches is faster, but depending on the size of your corpus you might run out of memory (in which case you should use smaller batches). At the time of writing the total number of strings is roughtly 700,000.}

\item{low_memory}{if TRUE (default) then data will be sorted in a way that tries to get a roughly equal number of string matches per batch, to prevent huge match tables (costing memory). If FALSE, data will be sorted in a way to get fewer unique words per batch, which can speed up matching, but can lead to a very unequal number of matches per batch.}
}
\details{
The resource needs to be downloaded first. For this you can use the download_resource function, which will (by default) download the resource into the tcorpus package folder.
}

