% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.r
\name{laplace}
\alias{laplace}
\title{Laplace (i.e. add constant) smoothing}
\usage{
laplace(freq, add = 0.5)
}
\arguments{
\item{freq}{A numeric vector of term frequencies (integers).}

\item{add}{The added value}
}
\value{
A numeric vector with the smoothed term proportions
}
\description{
Laplace (i.e. add constant) smoothing
}
