% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.r
\name{lda_fit}
\alias{lda_fit}
\title{Estimate a topic model using the LDA function of the topicmodels package}
\usage{
lda_fit(dtm, method = "Gibbs", K = 50, num.iterations = 500,
  alpha = 50/K, eta = 0.01, burnin = 250)
}
\arguments{
\item{dtm}{a document term matrix}

\item{method}{set method. see documentation for LDA function of the topicmodels package}

\item{K}{the number of clusters}

\item{num.iterations}{the number of iterations}

\item{alpha}{the alpha parameter}

\item{eta}{the eta parameter}

\item{burnin}{The number of burnin iterations}
}
\value{
A fitted LDA model
}
\description{
Estimate an LDA topic model using the LDA function from the topicmodels package.
The parameters other than dtm are simply passed to the sampler but provide a workable default.
See the description of that function for more information
}
