% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.r
\name{lda_plot_alltopics}
\alias{lda_plot_alltopics}
\title{Plot all topics}
\usage{
lda_plot_alltopics(m, time_var, category_var, path, date_interval = "day",
  value = "relative", create_index = T, ...)
}
\arguments{
\item{m}{The output of \code{\link{collapsed.gibbs.sampler}}}

\item{time_var}{A vector with time stamps (either numeric or Date class) of the same length and order of the documents (rows) in m$document_sums}

\item{category_var}{A vector with id values of the same length and order of the documents (rows) in m$document_sums}

\item{path}{The path for a folder where output will be saved}

\item{date_interval}{The interval for plotting the values over time. Can be: 'day', 'week', 'month' or 'year'}

\item{...}{Additional options passed to lda_plot_topic (and on to lda_plot_wordcloud)}
}
\value{
Nothing
}
\description{
Write plots for all topics with \code{\link{plot.topic}} in designated folder
}

