% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.r
\name{lda_plot_category}
\alias{lda_plot_category}
\title{Plots topic values per category}
\usage{
lda_plot_category(m, topic_nr, category_var, pct = F, value = "total",
  return.values = F)
}
\arguments{
\item{m}{The output of \code{\link{LDA}}}

\item{category_var}{A vector with id values of the same length and order of the documents (rows) in m$document_sums}

\item{pct}{Show topic values as percentages}

\item{value}{Show topic values as 'total', or as 'relative' to the attention for other topics}

\item{return.values}{Logical. If true, data that is plotted is returned as a data.frame}

\item{The}{index of the topic (1 to K)}
}
\value{
data.frame for plotted values
}
\description{
Plots the attention for a topic per category
}

