% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.r
\name{lda_plot_time}
\alias{lda_plot_time}
\title{Plots topic values over time}
\usage{
lda_plot_time(m, topic_nr, time_var, date_interval = "day", pct = F,
  value = "relative", return.values = F)
}
\arguments{
\item{m}{The output of \code{\link{LDA}}}

\item{time_var}{A vector with time stamps (either numeric or Date class) of the same length and order of the documents (rows) in m$document_sums}

\item{date_interval}{The interval for plotting the values over time. Can be: 'day', 'week', 'month' or 'year'}

\item{pct}{Show topic values as percentages}

\item{value}{Show topic values as 'total', or as 'relative' to the attention for other topics}

\item{return.values}{Logical. If true, data that is plotted is returned as a data.frame}

\item{The}{index of the topic (1 to K)}
}
\value{
data.frame for plotted values
}
\description{
Plots the attention for a topic over time
}

