% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.r
\name{lda_plot_topic}
\alias{lda_plot_topic}
\title{Plots topic wordcloud, and attention over time and per category}
\usage{
lda_plot_topic(m, topic_nr, time_var, category_var, date_interval = "day",
  pct = F, value = "relative", ...)
}
\arguments{
\item{m}{The output of \code{\link{collapsed.gibbs.sampler}}}

\item{time_var}{A vector with time stamps (either numeric or Date class) of the same length and order of the documents (rows) in m$document_sums}

\item{category_var}{A vector with id values of the same length and order of the documents (rows) in m$document_sums}

\item{date_interval}{The interval for plotting the values over time. Can be: 'day', 'week', 'month' or 'year'}

\item{pct}{Show topic values as percentages}

\item{value}{Show topic values as 'total', or as 'relative' to the attention for other topics}

\item{...}{Additional options passed to lda_plot_wordcloud}

\item{The}{index of the topic (1 to K)}
}
\value{
Nothing, just plots
}
\description{
Plots \code{\link{plot.wordcloud}}, \code{\link{plot.time}} and \code{\link{plot.category}}
}

