% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.r
\name{lda_plot_wordcloud}
\alias{lda_plot_wordcloud}
\title{Plot wordcloud for LDA topic}
\usage{
lda_plot_wordcloud(m, topic_nr, ...)
}
\arguments{
\item{m}{The output of \code{\link{LDA}}}

\item{topic_nr}{The index of the topic (1 to K)}

\item{...}{Additional options passed to dtm.wordcloud, e.g. freq.fun and pal}
}
\value{
Nothing, just plots
}
\description{
Plots a wordcloud of the top words per topic
}

