% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.r
\name{list_resources}
\alias{list_resources}
\title{List downloaded resources}
\usage{
list_resources(local_path = getOption("corpustools_resources", NULL))
}
\arguments{
\item{local_path}{Optionally, specify a path instead of using the corpustool directory}
}
\description{
Returns a vector with the filenames of downloaded resources.
}
