% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.r
\name{load_resource}
\alias{load_resource}
\title{Load one of the language resources that have been downloaded with download_resource().}
\usage{
load_resource(resource = c("jrc_names"),
  local_path = getOption("tcorpus_resources", NULL))
}
\arguments{
\item{resource}{A character string indicating which resource to download}

\item{local_path}{The path (without the filename) where the resource is stored. Please consult the documentation for download_resource() for more information regarding downloading and storing resources.}
}
\description{
Currently available resources:
\itemize{
   \item{JRC-Names: "JRC-Names is a highly multilingual named entity resource for person and organisation names. [...] JRC-Names is a by-product of the analysis of about 220,000 news reports per day by the Europe Media Monitor (EMM) family of applications." (https://ec.europa.eu/jrc/en/language-technologies/jrc-names)}
}
}

