% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_tcorpus.r
\name{merge_shards}
\alias{merge_shards}
\title{Merge tCorpus shards}
\usage{
merge_shards(shards)
}
\arguments{
\item{shards}{a named list or named arguments with tCorpus objects}
}
\value{
a tCorpus object
}
\description{
Similar to merge_tcorpora, but assumes that the tCorpora are identical in structure (same columns in the same order and same provenance)
}

