% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_tcorpus.r
\name{merge_tcorpora}
\alias{merge_tcorpora}
\title{Merge tCorpus objects}
\usage{
merge_tcorpora(..., keep_data = c("intersect", "all"),
  keep_meta = c("intersect", "all"))
}
\arguments{
\item{...}{either a named list or named arguments with tCorpus objects}

\item{keep_data}{if 'intersect', then only the token data columns that occur in all tCorpurs objects are kept}

\item{keep_meta}{if 'intersect', then only the document meta columns that occur in all tCorpurs objects are kept}
}
\value{
a tCorpus object
}
\description{
Merge tCorpus objects
}
