% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered_tcorpus.r
\name{open_shattered_tc}
\alias{open_shattered_tc}
\title{Open a shattered_tCorpus stored on disk}
\usage{
open_shattered_tc(path)
}
\arguments{
\item{path}{a path to a .tCorpus folder}
}
\description{
If a shattered_tCorpus has been created, this function can be used to read/open/load it by
}
\details{
For the sake of reference: a shattered_tCorpus is nothing more than an object that refers to a folder containing the tCorpus shards and index.
To create a shattered_tCorpus, consult the documentation ?shattered_tCorpus.
}
