% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.r
\name{order_meta}
\alias{order_meta}
\title{Add document meta to LDA output}
\usage{
order_meta(m, meta, match.by = "id")
}
\arguments{
\item{m}{The output of \code{\link{LDA}}}

\item{meta}{A data.frame with document meta. Has to contain a vector to match the document.ids}

\item{match.by}{The name of the vector in meta that matches the document.ids}
}
\value{
The LDA output appended with document meta
}
\description{
Add a dataframe containing document meta to the output (a list) of \code{\link{lda.collapsed.gibbs.sampler}}.
}

