% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_results.r
\name{plot.contextHits}
\alias{plot.contextHits}
\title{S3 plot for contextHits class}
\usage{
\method{plot}{contextHits}(object, min_weight = 0, backbone_alpha = NA,
  ...)
}
\arguments{
\item{object}{a contextHits object, as returned by \link{search_contexts}}

\item{...}{not used}
}
\description{
S3 plot for contextHits class
}
\examples{
tc = create_tcorpus(sotu_texts, doc_column='id')
hits = search_contexts(tc, c('War# war* OR army OR bomb*','Terrorism# terroris*',
                              Economy# econom* OR bank*','Education# educat* OR school*'))
plot(hits)
}
