% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_words.r
\name{plot.featureAssociations}
\alias{plot.featureAssociations}
\title{visualize feature associations}
\usage{
\method{plot}{featureAssociations}(x, n = 25, size = c("chi2", "freq",
  "ratio"), ...)
}
\arguments{
\item{x}{a featureAssociations object, created with the \link{feature_associations} function}

\item{n}{the number of words in the plot}

\item{size}{use "freq", "chi2" or "ratio" for determining the size of words}

\item{...}{additional arguments passed to dtm_wordcloud}
}
\description{
visualize feature associations
}
\examples{
## as example, compare SOTU paragraphs about taxes to rest
tc = create_tcorpus(sotu_texts[1:100,], doc_column = 'id')
comp = tc$compare_subset('token', query_x = 'tax*')

\donttest{
plot(comp, balance=T)
plot(comp, mode = 'ratio_x')
plot(comp, mode = 'ratio_y')
}
}
