% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_words.r, R/semnet_plotting.r
\name{plotWords}
\alias{plotWords}
\title{Plot a wordcloud with words ordered and coloured according to a dimension (x)}
\usage{
plotWords(x, y = NULL, words, wordfreq = rep(1, length(x)), xlab = "",
  ylab = "", yaxt = "n", scale = 2, random.y = F, xlim = NULL,
  ylim = NULL, col = NULL, ...)

plotWords(x, y = NULL, words, wordfreq = rep(1, length(x)), xlab = "",
  ylab = "", yaxt = "n", scale = 2, random.y = F, xlim = NULL,
  ylim = NULL, col = NULL, ...)
}
\arguments{
\item{x}{The (approximate) x positions of the words}

\item{y}{The (approximate) y positions of the words}

\item{words}{A character vector with the words to plot}

\item{wordfreq}{The frequency of the words, defaulting to 1}

\item{x}{The (approximate) x positions of the words}

\item{y}{The (approximate) y positions of the words}

\item{words}{A character vector with the words to plot}

\item{wordfreq}{The frequency of the words, defaulting to 1}
}
\value{
nothing
}
\description{
Plot a wordcloud with words ordered and coloured according to a dimension (x)

Plot a wordcloud with words ordered and coloured according to a dimension (x)
}

