% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.r
\name{prepare_time_var}
\alias{prepare_time_var}
\title{Change date object to date_interval}
\usage{
prepare_time_var(time_var, date_interval)
}
\arguments{
\item{time_var}{A vector of Date values}

\item{date_interval}{The desired date_interval ('day','week','month', or 'year')}
}
\value{
A vector of Date values
}
\description{
Change date object to date_interval
}

