% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.r
\name{preprocess_feature}
\alias{preprocess_feature}
\title{Preprocess feature}
\usage{
preprocess_feature(tc, column, new_column, lowercase = T, ngrams = 1,
  ngram_context = c("document", "sentence"), as_ascii = F,
  remove_punctuation = T, remove_stopwords = F, use_stemming = F,
  language = "english")
}
\description{
Preprocess feature
}

