% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.r
\name{preprocess_words}
\alias{preprocess_words}
\title{Preprocess words in a character vector}
\usage{
preprocess_words(x, context = NULL, language = "english",
  use_stemming = F, lowercase = T, ngrams = 1, replace_whitespace = T,
  as_ascii = F, remove_punctuation = T, remove_stopwords = F)
}
\arguments{
\item{x}{A character vector in which each element is a word (i.e. a tokenized text)}

\item{context}{Optionally, a character vector of the same length as x, specifying the context of word (e.g., document, sentence). Has to be given if ngram > 1}

\item{language}{The language used for stemming and removing stopwords}

\item{use_stemming}{Logical, use stemming. (Make sure the specify the right language!)}

\item{lowercase}{Logical, make word lowercase}

\item{ngrams}{A number, specifying the number of words per ngram. Default is unigrams (1).}

\item{replace_whitespace}{Logical. If TRUE, all whitespace is replaced by underscores}

\item{as_ascii}{Logical. If TRUE, words will be forced to ascii}

\item{remove_punctuation}{Logical. if TRUE, punctuation is removed}

\item{remove_stopwords}{Logical. If TRUE, stopwords are removed (Make sure to specify the right language!)}
}
\description{
Preprocess words in a character vector
}

