% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_results.r
\name{print.contextHits}
\alias{print.contextHits}
\title{S3 print for contextHits class}
\usage{
print.contextHits(x, ...)
}
\arguments{
\item{x}{a contextHits object, as returned by \link{tCorpus$search_contexts}}

\item{...}{not used}
}
\description{
S3 print for contextHits class
}

