% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_results.r
\name{print.featureHits}
\alias{print.featureHits}
\title{S3 print for featureHits class}
\usage{
print.featureHits(x, ...)
}
\arguments{
\item{x}{a featureHits object, as returned by \link{tCorpus$search_features}}

\item{...}{not used}
}
\description{
S3 print for featureHits class
}
