% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered_tcorpus.r
\name{print.shattered_tCorpus}
\alias{print.shattered_tCorpus}
\title{S3 print for shattered_tCorpus class}
\usage{
print.shattered_tCorpus(x, ...)
}
\arguments{
\item{x}{a shattered_tCorpus object}

\item{...}{not used}
}
\description{
S3 print for shattered_tCorpus class
}
