% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcorpus.r
\name{print.tCorpus}
\alias{print.tCorpus}
\title{S3 print for tCorpus class}
\usage{
\method{print}{tCorpus}(x, ...)
}
\arguments{
\item{x}{a tCorpus object}

\item{...}{not used}
}
\description{
S3 print for tCorpus class
}
