% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{queryTermMatches}
\alias{queryTermMatches}
\title{Report number of unique terms that matches each query}
\usage{
queryTermMatches(terms, queries)
}
\arguments{
\item{terms}{a character vector of (unique) terms}

\item{queries}{a dataframe with queries as used in the feature_queries en code_queries functions.}
}
\description{
This function is usefull to test whether certain queries match a high number of unique terms, thus possibly causing memory issues.
}

