% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcorpus.r
\name{recode_meta_column}
\alias{recode_meta_column}
\title{Recode a document meta column in a tCorpus}
\usage{
recode_meta_column(tc, column, new_value, i = NULL, old_value = NULL)
}
\arguments{
\item{column}{the name of the document meta column}

\item{new_value}{the new value}

\item{i}{an index for which values to replace}

\item{old_value}{a vector containing one or more old values to replace}
}
\description{
Recode a document meta column in a tCorpus
}

