% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shatter_tcorpus.r
\name{redistribute_shards}
\alias{redistribute_shards}
\title{Redistribute shards in a shattered tCorpus}
\usage{
redistribute_shards(tc_index, tokens_per_shard = 10000)
}
\arguments{
\item{tokens_per_shard}{}
}
\description{
For good performance it's generally best to keep shards of moderate size. Not too small to prevent tedious merging and looping, and not too big to prevent memory problems. This function redistributes the shards (within categories) to approximate the specified tokens_per_shard
}

