% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicbrowser.r
\name{render_article}
\alias{render_article}
\title{Render a single article}
\usage{
render_article(doc, info, maxwords = NULL)
}
\arguments{
\item{info}{the cluster_info object (list)}

\item{doc:}{the document id (dtm rowname) to render}
}
\description{
This function does not return the html, but rather cats it directly
So you might wish to use wrap_html or sink to capture the output
}

