% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicbrowser.r
\name{render_overview}
\alias{render_overview}
\title{Render the index page}
\usage{
render_overview(info)
}
\arguments{
\item{info}{the cluster_info object (list)}
}
\description{
This function does not return the html, but rather cats it directly
So you might wish to use wrap_html or sink to capture the output
}

