% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicbrowser.r
\name{render_topic}
\alias{render_topic}
\title{Render a single topic}
\usage{
render_topic(topic_id, info, nmaxdoc = 10)
}
\arguments{
\item{info}{the cluster_info object (list)}

\item{topic_id:}{the topic id to render}
}
\description{
This function does not return the html, but rather cats it directly
So you might wish to use wrap_html or sink to capture the output
}

