% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_results.r
\name{reportKWIC}
\alias{reportKWIC}
\title{Title}
\usage{
reportKWIC(tc, hits = NULL, token_i = NULL, nwords = 10, nsample = 10,
  random_sample = T, doc.col = getOption("doc.col", "doc_id"),
  position.col = getOption("position.col", "position"),
  feature.col = getOption("wor.col", "word"))
}
\arguments{
\item{feature.col}{}
}
\description{
Title
}

