% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_results.r
\name{reportXnotY}
\alias{reportXnotY}
\title{Title}
\usage{
reportXnotY(tokens, hits.x, hits.y, doc.col = getOption("doc.col", "doc_id"),
  position.col = getOption("position.col", "position"),
  feature.col = getOption("wor.col", "word"), ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}

