% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{require_package}
\alias{require_package}
\title{Check if package with given version exists}
\usage{
require_package(package, min_version = NULL)
}
\arguments{
\item{package}{The name of the package}

\item{min_version}{The minimum version}
}
\value{
An error if package does not exist
}
\description{
Check if package with given version exists
}
