% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcorpus.r
\name{reset_feature_index}
\alias{reset_feature_index}
\title{Reset feature index}
\usage{
reset_feature_index(tc)
}
\arguments{
\item{tc}{}
}
\description{
Reset the feature_index, using the . This is for instance necessary after filtering the data without keeping the data
}

