% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.r
\name{resources_path}
\alias{resources_path}
\title{Get name of the resources location}
\usage{
resources_path(local_path = getOption("corpustools_resources", NULL))
}
\arguments{
\item{local_path}{Optionally, specify a path instead of using the corpustool directory}
}
\description{
Get name of the resources location
}
