% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_contexts.r
\name{search_contexts}
\alias{search_contexts}
\title{Title}
\usage{
search_contexts(tc, query, code = NULL, feature = "word",
  context_level = c("document", "sentence"), verbose = F)
}
\arguments{
\item{context_level}{}
}
\description{
Title
}

