% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_features.r
\name{search_recode}
\alias{search_recode}
\title{Recode features in a tCorpus based on a search string}
\usage{
search_recode(tc, feature, new_value, keyword, condition = NA,
  condition_once = F, keyword_filter = NULL)
}
\arguments{
\item{keyword_filter}{}
}
\description{
Search features (see documentation for search_features function for instructions) and replace features with a new value
}

