% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_index.r
\name{set_feature_index}
\alias{set_feature_index}
\title{Create a feature index within a tCorpus}
\usage{
set_feature_index(tc, feature = "word", context_level = c("document",
  "sentence"), max_window_size = 100)
}
\arguments{
\item{max_window_size}{}
}
\description{
Create a feature index within a tCorpus
}

