% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.r
\name{set_resources_path}
\alias{set_resources_path}
\title{Specifiy the path where the downloaded resources are stored}
\usage{
set_resources_path(path = NULL)
}
\arguments{
\item{path}{The path (without the filename) where the resources are stored. If NULL, the file will be stored in the location of the tcorpus package. if tempdir(), the file is stored in a directory that is cleaned up when the R session ends.}
}
\description{
By default, tcorpus stores the resources within the tcorpus directory. With set_resources_path you can change this default to a custom path.
}
