% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.r
\name{sgt}
\alias{sgt}
\title{Simple Good Turing smoothing}
\usage{
sgt(freq)
}
\arguments{
\item{freq}{A numeric vector of term frequencies (integers).}
}
\value{
A numeric vector with the smoothed term proportions
}
\description{
Implementation of the Simple Good Turing smoothing proposed in: Gale, W. A., \& Sampson, G. (1995). Good turing frequency estimation without tears. Journal of Quantitative Linguistics, 2(3), 217-237.
}
