% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpipe.r
\name{show_udpipe_models}
\alias{show_udpipe_models}
\title{Show the names of udpipe models}
\usage{
show_udpipe_models()
}
\value{
a data.frame
}
\description{
Returns a data.table with the language, treebank and udpipe_model name.
Uses the default model repository provided by the udpipe package (\code{\link[udpipe]{udpipe_download_model}}).
For more information about udpipe and performance benchmarks of the UD models, see the
GitHub page of the \href{https://github.com/bnosac/udpipe}{udpipe package}.
}
\examples{
show_udpipe_models()
}
