% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{stopwords_list}
\alias{stopwords_list}
\title{Basic stopword lists}
\format{A named list, with names matching the languages used by SnowballC}
\usage{
data(stopwords_list)
}
\description{
Basic stopword lists
}
\keyword{datasets}
