% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcorpus.r
\name{subset.tCorpus}
\alias{subset.tCorpus}
\title{Subset a tCorpus}
\usage{
\method{subset}{tCorpus}(tc, subset = NULL, subset_meta = NULL)
}
\arguments{
\item{tc}{tCorpus object}

\item{subset}{}
}
\description{
Subset a tCorpus
}

