% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.r
\name{subset_feature_fun}
\alias{subset_feature_fun}
\title{Filter features}
\usage{
subset_feature_fun(tc, column, new_column, subset, inverse = F)
}
\arguments{
\item{filter}{}
}
\description{
Similar to subsetting, but instead of deleting rows, the rows for the specified column as set to NA. This way the vocabulary can be reduced while still beign able to bring results of analyses back to the full text
}

