% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_contexts.r
\name{subset_query}
\alias{subset_query}
\title{Title}
\usage{
subset_query(tc, query, code = NULL, feature = "word",
  context_level = c("document", "sentence"))
}
\arguments{
\item{context_level}{}
}
\description{
Title
}

