% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_tcorpus.r
\name{subset_query_window}
\alias{subset_query_window}
\title{Subset a tcorpus for a window around the results of search_features()}
\usage{
subset_query_window(tc, window, keyword = NA, condition = NA,
  queries = NULL, feature = "word", condition_once = F,
  subset_tokens = NA, subset_meta = NA, verbose = F)
}
\arguments{
\item{verbose}{}
}
\value{
a tCorpus object
}
\description{
This function works like the search_features() function, but instead of returning a data.frame with the search results, it returns a subset of the tcorpus with the search results, and a specified window arround the results.
}

