% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_tcorpus.r
\name{subset_window}
\alias{subset_window}
\title{subset a tCorpus to all tokens that occur within the given window of a given set of indices}
\usage{
subset_window(tc, i, window, context_level = c("document", "sentence"))
}
\arguments{
\item{tc}{a tCorpus object}

\item{i}{the indices}

\item{window}{an integer specifying the window size}

\item{context_level}{}
}
\description{
subset a tCorpus to all tokens that occur within the given window of a given set of indices
}

