% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_results.r
\name{summary.contextHits}
\alias{summary.contextHits}
\title{S3 summary for contextHits class}
\usage{
\method{summary}{contextHits}(object, ...)
}
\arguments{
\item{object}{a contextHits object, as returned by \link{tCorpus$search_contexts}}

\item{...}{not used}
}
\description{
S3 summary for contextHits class
}
