% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_results.r
\name{summary.featureHits}
\alias{summary.featureHits}
\title{S3 summary for featureHits class}
\usage{
\method{summary}{featureHits}(object, ...)
}
\arguments{
\item{object}{a featureHits object, as returned by \link{tCorpus$search_features}}

\item{...}{not used}
}
\description{
S3 summary for featureHits class
}
