% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered_tcorpus.r
\name{summary.shattered_tCorpus}
\alias{summary.shattered_tCorpus}
\title{S3 summary for shattered_tCorpus class}
\usage{
summary.shattered_tCorpus(object, ...)
}
\arguments{
\item{object}{a shattered_tCorpus object}

\item{...}{not used}
}
\description{
S3 summary for shattered_tCorpus class
}

