% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_tcorpus.r
\name{summary.tCorpus}
\alias{summary.tCorpus}
\title{Summary of a tCorpus object}
\usage{
\method{summary}{tCorpus}(object, ...)
}
\arguments{
\item{object}{A tCorpus object}

\item{...}{not used}
}
\description{
Summary of a tCorpus object
}
\examples{
tc = create_tcorpus(c('First text', 'Second text'))
summary(tc)
}
