% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicbrowser.r
\name{tagTokens}
\alias{tagTokens}
\title{Add title, href, and class to tokens}
\usage{
tagTokens(tokens, topics)
}
\arguments{
\item{tokens:}{a vector of words}

\item{topics:}{a vector of topics, can include NA's for words without a topic}
}
\value{
a vector of raw html per token
}
\description{
Add title, href, and class to tokens
}

