% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tc_sotu_udpipe}
\alias{tc_sotu_udpipe}
\title{A tCorpus with a small sample of sotu paragraphs parsed with udpipe}
\format{
data.frame
}
\usage{
data(tc_sotu_udpipe)
}
\description{
A tCorpus with a small sample of sotu paragraphs parsed with udpipe
}
\keyword{datasets}
