% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsyntax_convenience.r
\name{tc_syntax_reader}
\alias{tc_syntax_reader}
\title{Create a full text browser with highlighted rsyntax annotations}
\usage{
tc_syntax_reader(
  tc,
  annotation = NULL,
  value = NULL,
  value2 = NULL,
  doc_ids = NULL,
  token_col = "token",
  n = 500,
  select = c("first", "random"),
  meta_cols = NULL,
  seed = NA,
  header = "",
  subheader = NULL,
  ...
)
}
\arguments{
\item{tc}{a tCorpus}

\item{annotation}{Optionally, a column with an rsyntax annotation, to add boxes around the annotated nodes.}

\item{value}{A character vector with values used in the annotation column. If used, only these values are fully colored, and the other (non NA) values only have border colors. For example, if annotation distinguishes 'source', 'verb' and 'quote', then value = 'source' will highlight the source tokens and only color the borders of the 'verb' and 'quote' tokens'}

\item{value2}{An extension of value. Normally, all (non NA) values that are not in value will be given border colors. With value2 you can specify which specific values to give border colors}

\item{doc_ids}{A vector with document ids to view}

\item{token_col}{The name of the column in tc$tokens that contain the token text}

\item{n}{Only n of the results are printed (to prevent accidentally making huge browsers).}

\item{select}{If n is smaller than the number of documents in tc, select determines how the n documents are selected}

\item{meta_cols}{A character vector with names of columns in tc$meta, used to only show the selected columns}

\item{seed}{If select is "random", seed can be used to set a random seed. After sampling the seed is re-initialized with set.seed(NULL).}

\item{header}{Optionally, give a custom header}

\item{subheader}{Optionally, give a custom subheader (overwriting some of the default information about n and sampling)}

\item{...}{Arguments passed to \code{\link[rsyntax]{syntax_reader}}}
}
\value{

}
\description{
A wrapper for the \code{\link[rsyntax]{syntax_reader}} function, that can be used directly on a tCorpus, and adopts some of the features from \code{\link{browse_texts}}
Pro tip for fixing/improving tqueries: if you hover over the highlighted tokens, you see the annotation id. If you labeled your tqueries you can thus easily see which query was applied, and you can look up the id in your tokens data (annotation ids are unique at the corpus level)
}
\examples{
tc = tc_sotu_udpipe$copy()

tc_syntax_reader(tc)

tc$annotate_clauses(column = 'clauses')
tc_syntax_reader(tc, annotation='clauses', value='subject',n=2)
}
