% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_dictionary.r
\name{tCorpus$code_dictionary}
\alias{tCorpus$code_dictionary}
\alias{code_dictionary}
\title{Dictionary lookup}
\arguments{
\item{dict}{A data.frame with the columns "string" and "id". The text in "string" will be matched to the tokens, and "id" will be the id given to tokens that matched}

\item{regex_sep}{A regular expression for splitting multiple terms in dict$string. By default whitespace.}

\item{token_col}{The feature in tc that contains the token text.}

\item{column}{The name of the new column in tc$tokens. Default is "code"}

\item{case_sensitive}{logical, should lookup be case sensitive?}

\item{batchsize}{Very large dictionaries will be matched in batches to prevent memory issues.}

\item{flatten_colloc}{If true, collocations in the tokens (rows in tc$tokens) will be considered separate words. For example, "President_Obama" will be split to "president" "obama", so that "president obama" in the dictionary matches correctly.}

\item{ascii}{If true, convert text to ascii before matching}

\item{low_memory}{if true, use slower but more memory efficient algorithm}

\item{verbose}{If true, report progress}
}
\value{
A vector with the id value (taken from dict$id) for each row in tc$tokens
}
\description{
Add a column to the token data that contains a code (the query label) for tokens that match the dictionary

\strong{Usage:}

## R6 method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{code_dictionary(...)}
}
\examples{
dict = data.frame(string = c('this is', 'a', 'test'), id=1:3)
tc = create_tcorpus(c('this is a test','This town is not big enough for a test'))
tc$code_dictionary(dict)
print(tc$tokens)
}
