% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_features.r
\name{tCorpus$code_features}
\alias{tCorpus$code_features}
\alias{code_features}
\title{Code features in a tCorpus based on a search string}
\arguments{
\item{query}{A character string that is a query. See details for available query operators and modifiers. Can be multiple queries (as a vector), in which case it is recommended to also specifiy the code argument, to label results.}

\item{code}{The code given to the tokens that match the query (usefull when looking for multiple queries). Can also put code label in query with # (see details)}

\item{feature}{The name of the feature column within which to search.}

\item{column}{The name of the column that is added to the data}

\item{...}{Additional search_features parameters. See \link{tCorpus$search_features}}
}
\description{
Add a column to the token data that contains a code (the query label) for tokens that match the query (see \link{tCorpus$search_features}).
}
\section{Usage}{

## R6 method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{
code_features(query, code=NULL, feature='token', column='code', ...)
}
}

\examples{
tc = create_tcorpus('Anna and Bob are secretive')

tc$code_features(c("actors# anna bob", "associations# secretive"))
tc$get()
}
