% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$data}
\alias{tCorpus$data}
\alias{data.tCorpus}
\alias{tCorpus$meta}
\alias{meta.tCorpus}
\title{Access specific column from the data or meta data.}
\description{
The $data and $meta fields can be used to extract the token and meta data.

In addition, it can be used to modify the data (with some limitations) using common assignment syntax.
However, for efficiency it is recommended to use the accessor methods (i.e. the set*, get*, subset* and delete* methods)

Limitations for modifying by assignment:
- subsetting or appending the data is not possible. (use the subset function and merge_tcorpua)
- data.table operations that change the data.table by reference are not possible. Note that the set and set_meta functions do use assignment by reference.
- the returned data.table is a copy of the data.table within the corpus. This copy is an actual copy (the copy-on-modify mechanic doesn't work here) which can eat up memory if. Accordingly, if you only want to get 1 or several columns, it is more efficient to use the get() and get_meta() methods, which only copy the selected columns.
}
\section{Usage}{

## R6 method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{get}
\preformatted{meta}
}

