% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$feature_stats}
\alias{feature_stats.tCorpus}
\alias{tCorpus$feature_stats}
\title{Feature statistics}
\arguments{
\item{feature}{The name of the feature}

\item{sent_freq}{If True, include sentence frequency (only if sentence information is available).}
}
\description{
Compute a number of useful statistics for features: term frequency, idf, etc.
}
\section{Usage}{

## R6 method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{feature_stats(feature, sent_freq=F)}
}

