% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$filter}
\alias{filter.tCorpus}
\alias{tCorpus$filter}
\title{Filter feature}
\usage{
## R6 method for class tCorpus
\code{tCorpus$filter(column, new_column, filter, clone=self$clone_on_change)}
}
\arguments{
\item{column}{the column containing the feature to be used as the input}

\item{new_column}{the column to save the filtered feature. Can be a new column or overwrite an existing one.}

\item{filter}{logical expression indicating rows to keep in the tokens data. i.e. rows for which the logical expression is FALSE will be set to NA.}

\item{clone}{If TRUE, the method returns a new tCorpus object. This is the normal R way of doing things. Alternatively, the tCorpus can be used as a reference class object by setting clone to FALSE, or setting tCorpus$clone_on_change to FALSE to use this globally. Please consult the general documentation for tCorpus (?tCorpus) for a more detailed explanation.}
}
\description{
Similar to using \link{tCorpus$subset}, but instead of deleting rows it only sets rows for a specified feature to NA. This can be very convenient, because it enables only a selection of features to be used in an analysis (e.g., a topic model) but maintaining the context of the full article, so that results can be viewed in this context (e.g., a topic browser).

Just as in subset, it is easy to use objects and functions in the filter, including the special functions for using term frequency statistics (see documentation for \link{tCorpus$subset}).
}

