% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$kwic}
\alias{kwic.tCorpus}
\alias{tCorpus$kwic}
\title{Get keyword-in-context (KWIC) strings}
\usage{
## R6 method for class tCorpus
}
\arguments{
\item{hits}{results of feature search. see \link{tCorpus$search_features}.}

\item{i}{instead of the hits argument, you can give the indices of features directly.}

\item{keyword}{instead of using the hits or i arguments, a search string can be given directly. Note that this simply a convenient shorthand for first creating a hits object with \link{tCorpus$search_features}. If a keyword is given, then the ... argument is used to pass other arguments to \link{tCorpus$search_features}.}

\item{code}{if 'i' or 'keyword' is used, the code argument can be used to add a code label. Should be a vector of the same length that gives the code for each i or keyword, or a vector of length 1 for a single label.}

\item{nwords}{an integers specifying the size of the context, i.e. the number of words left and right of the keyword.}

\item{nsample}{optionally, get a random sample of the keywords/features. If multiple codes are used, the sample is drawn for each code individually.}

\item{output_feature}{the feature column that is used to make the KWIC.}

\item{context_level}{Select the maxium context (document or sentence).}

\item{prettypaste}{If TRUE, tries to reconstruct the text in a natural way (e.g. no space before a dot). Otherwise, all features are separated by a whitespace.}

\item{kw_tag}{a character vector of length 2, that gives the symbols before (first value) and after (second value) the keyword in the KWIC string. Can for instance be used to prepare KWIC with format tags for highlighting.}

\item{...}{See \link{tCorpus$search_features} for the query parameters}
}
\description{
Create a data.frame with keyword-in-context strings for given indices (i), search results (hits) or search strings (keyword).
}

