% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_text.r
\name{tCorpus$read_text}
\alias{tCorpus$read_text}
\alias{read_text.tCorpus}
\title{Print tokens as text}
\arguments{
\item{doc_id}{The doc_ids of the documents to be printed.}

\item{column}{The name of the column from which the text is printed.}

\item{meta_columns}{The meta data that is printed at the top of each text.}
}
\description{
Print tokens as text
}
\section{Usage}{

## R6 method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{
read_text(doc_id, column='token', meta_columns = self$meta_names)}
}

