% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$search_contexts}
\alias{search_contexts.tCorpus}
\alias{tCorpus$search_contexts}
\title{Search for documents or sentences using Boolean queries}
\usage{
## R6 method for class tCorpus
\code{tCorpus$search_contexts(query, code=NULL, feature='word', context_level=c('document','sentence'), verbose=F)}
}
\arguments{
\item{query}{A character string that is a query. See details for available query operators and modifiers. Can be multiple queries (as a vector), in which case it is recommended to also specifiy the code argument, to label results.}

\item{code}{If given, used as a label for the results of the query. Especially usefull if multiple queries are used.}

\item{feature}{The name of the feature column}

\item{context_level}{Select whether the queries should occur within while "documents" or specific "sentences". Returns results at the specified level.}

\item{verbose}{}
}
\description{
Search for documents or sentences using Boolean queries
}

