% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$search_recode}
\alias{search_recode.tCorpus}
\alias{tCorpus$search_recode}
\title{Recode features in a tCorpus based on a search string}
\usage{
## R6 method for class tCorpus
}
\arguments{
\item{feature}{The feature in which to search}

\item{new_value}{the character string with which all features that are found are replaced}

\item{...}{See \link{tCorpus$search_features} for the query parameters}

\item{copy}{If TRUE, the method returns a new tCorpus object. This is the normal R way of doing things. Alternatively, the tCorpus can be used as a reference class object by setting copy to FALSE, or setting tCorpus$copy_on_modify to FALSE to use this globally. Please consult the general documentation for tCorpus (?tCorpus) for a more detailed explanation.}
}
\description{
Search features (see \link{tCorpus$search_features}) and replace features with a new value
}

