% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$set}
\alias{set.tCorpus}
\alias{set_meta.tCorpus}
\alias{tCorpus$set}
\alias{tCorpus$set_meta}
\title{Modify the token and meta data.tables of a tCorpus}
\usage{
## R6 method for class tCorpus

## R6 method for class tCorpus.
}
\arguments{
\item{column}{Name of a new column (to create) or existing column (to transform)}

\item{value}{A vector of the same length as the number of rows in the data. Note that if a subset is used, the length of value should be the same as the length of the subset (the TRUE cases of the subset expression) or a single value.}

\item{subset}{logical expression indicating rows to keep in the tokens data or meta data}

\item{copy}{If TRUE, the method returns a new tCorpus object. This is the normal R way of doing things. Alternatively, the tCorpus can be used as a reference class object by setting copy to FALSE, or setting tCorpus$copy_on_modify to FALSE to use this globally. Please consult the general documentation for tCorpus (?tCorpus) for a more detailed explanation.}

\item{safe}{If TRUE, you are unable to modify the position columns (doc_id, sent_i, word_i), which is very likely to break the tCorpus. Only set to FALSE if you know exactly what you're doing.}
}
\description{
Modify the token/meta data.table by setting the values of one (existing or new) column. This is less flexible than within data or transform data, but it has the advantage of allowing columns to be selected as a string, which makes it convenient for modifying the tCorpus from within function. The subset argument can be used to modify only subsets of columns, and can be a logical vector (select TRUE rows), numeric vector (indices of TRUE rows) or logical expression (e.g. pos == 'noun'). If A new column is made whie using a subset, then the rows outside of the selection are set to NA.

Change column names in tCorpus data
}

