% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$set_colname}
\alias{tCorpus$set_colname}
\alias{set_colname.tCorpus}
\alias{tCorpus$set_meta_colname}
\alias{set_meta_colname.tCorpus}
\title{Change column names of data and meta data}
\arguments{
\item{oldname}{the current/old column name}

\item{newname}{the new column name}
}
\description{
Change column names of data and meta data
}
\section{Usage}{

## R6 method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{set_colname(oldname, newname)}
\preformatted{set_meta_colname(oldname, newname)}
}

