% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$set_special}
\alias{tCorpus$set_special}
\alias{set_special}
\title{Designate column as columns with special meaning (token, lemma, POS, relation, parent)}
\arguments{
\item{token}{Name of the column that will be designated as the token, and renamed to 'token'}

\item{lemma}{Name of the column that will be designated as the lemma of the token, and renamed to 'lemma'}

\item{pos}{Name of the column that will be designated as the part-of-speech tag of the token, and renamed to 'POS'}

\item{relation}{Name of the column that will be designated as the dependency relation of the token to its parent, and renamed to 'relation'}

\item{parent}{Name of the column that will be designated as the parent of the token, and renamed to 'parent'}
}
\description{
\strong{Usage:}
}
\details{
## R6 method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{set_special(token=NULL, lemma=NULL, POS=NULL, relation=NULL, parent=NULL)}
}
