% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$subset_query}
\alias{tCorpus$subset_query}
\alias{subset_query.tCorpus}
\title{Subset tCorpus token data using a query}
\arguments{
\item{query}{A character string that is a query. See \link{tCorpus$search_contexts} for query syntax.}

\item{feature}{The name of the feature columns on which the query is used.}

\item{context_level}{Select whether the query and subset are performed at the document or sentence level.}
}
\description{
A convenience function that searches for contexts (documents, sentences), and uses the results to \link[=tCorpus$search_contexts]{subset} the tCorpus token data.

See the documentation for \link[=tCorpus$search_contexts]{subset} for an explanation of the query language.
}
\section{Usage}{

## R6 method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{subset_query(query, feature = 'token', context_level = c('document','sentence'))}
}

