% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$transform}
\alias{tCorpus$transform}
\alias{tCorpus$transform_meta}
\alias{transform.tCOrpus}
\alias{transform_meta.tCorpus}
\title{Modify the token and meta data.tables of a tCorpus}
\usage{
## R6 method for class tCorpus
\code{tCorpus$transform(..., clone=self$clone_on_change, safe=T)}
\code{tCorpus$transform_meta(..., clone=self$clone_on_change, safe=T)}
}
\arguments{
\item{...}{name-value pairs where the name indicates a new or existing column in the token data, to which the value is set.}

\item{clone}{If TRUE, the method returns a new tCorpus object. This is the normal R way of doing things. Alternatively, the tCorpus can be used as a reference class object by setting clone to FALSE, or setting tCorpus$clone_on_change to FALSE to use this globally. Please consult the general documentation for tCorpus (?tCorpus) for a more detailed explanation.}

\item{safe}{If TRUE, you are unable to modify the position columns (doc_id, sent_i, word_i), which is very likely to break the tCorpus. Only set to FALSE if you know exactly what you're doing.}
}
\description{
Modify the token/meta data.table using the \link{transform.data.table} function. Arguments are name-value pairs where the name indicates a new or existing column in the token data, to which the value is set. The value can be specified quite flexibly, and can directly call columns in the tokens data by name. For example, we can make a new column "word_low" by using the "word" column and making it lowercase: $transform_data(word_low = tolower(word)). Note that this can also be used to delete columns by setting their value to NULL: $transform_data(word_low = NULL).
}

