% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpipe_coref.r
\name{tCorpus$udpipe_coref}
\alias{tCorpus$udpipe_coref}
\alias{udpipe_coref}
\title{Perform rule based coreference resolution}
\arguments{
\item{keep_feats}{If TRUE, keep the feature columns created for the coref resolution}
}
\value{
a tCorpus
}
\description{
This is an experimental implementation of coreference resolution, aimed at coreferences of nouns and pronouns.
}
\examples{
tc = tc_sotu_udpipe$copy()
tc$udpipe_quotes()
\dontrun{
 tc$udpipe_coref()
 tc$tokens[!is.na(tc$tokens$coref_id),]
}
}
