% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcorpus.r
\docType{class}
\name{tCorpus-class}
\alias{tCorpus}
\alias{tCorpus-class}
\title{tCorpus; a corpus for tokenized texts}
\description{
tCorpus; a corpus for tokenized texts
}
\section{Slots}{

\describe{
\item{\code{data}}{data.table.}

\item{\code{meta}}{data.table.}

\item{\code{feature_index}}{data.table.}

\item{\code{p}}{list.}
}}

