% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_overview.r
\name{tCorpus}
\alias{tCorpus}
\alias{tcorpus}
\title{tCorpus: a corpus class for tokenized texts}
\description{
tCorpus: a corpus class for tokenized texts
}
\section{Working with the tCorpus}{

The primary goal of the tCorpus is to facilitate various corpus analysis techniques. The documentation for currently implemented techniques can be reached through the following links.
\tabular{ll}{
  \link[=tCorpus_create]{Create a tCorpus} \tab Functions for creating a tCorpus object \cr
  \link[=tCorpus_data]{Manage tCorpus data} \tab Methods for viewing, modifying and subsetting tCorpus data \cr
  \link[=tCorpus_features]{Features} \tab Preprocessing, filtering and analyzing features \cr
  \link[=tCorpus_querying]{Using search strings} \tab Use Boolean queries to analyze the tCorpus \cr
  \link[=tCorpus_semnet]{Co-occurrence networks}  \tab Feature co-occurrence based semantic network analysis \cr
  \link[=tCorpus_compare]{Corpus comparison} \tab Compare corpora \cr
  \link[=tCorpus_topmod]{Topic modeling}  \tab Create and visualize topic models \cr
  \link[=tCorpus_docsim]{Document similarity} \tab Calculate document similarity \cr
  \link[=tCorpus_resources]{Language resources} \tab An overview of external language resource plugins
}
}

