% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_information.r
\name{tcorpus_compare}
\alias{tcorpus_compare}
\title{Title}
\usage{
tcorpus_compare(tc_x, tc_y, feature, context_level = c("document",
  "sentence"), smooth = 1, min.over = NULL, min.chi = NULL,
  x_is_subset = F)
}
\arguments{
\item{x_is_subset}{}
}
\description{
Title
}

