% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_overview.r
\name{tCorpus_create}
\alias{tCorpus_create}
\title{Creating a tCorpus}
\description{
\link[=tCorpus]{(back to overview)}
}
\details{
\strong{Create a tCorpus}
\tabular{ll}{
  \link[=create_tcorpus]{create_tcorpus()} \tab Create a tCorpus from raw text input \cr
  \link[=tokens_to_tcorpus]{tokens_to_tcorpus()} \tab Create a tCorpus from a data.frame of already tokenized texts
}
}
