% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_overview.r
\name{tCorpus_data}
\alias{tCorpus_data}
\title{Methods for viewing, modifying and subsetting tCorpus data}
\description{
\link[=tCorpus]{(back to overview)}
}
\details{
\strong{Fields}
\tabular{ll}{
  $n \tab The number of tokens (i.e. rows in the data) \cr
  $n_meta \tab The number of documents (i.e. rows in the document meta data) \cr
  $names \tab The names of the token data columns \cr
  $names_meta \tab The names of the document meta data columns
}

\strong{Get data}
\tabular{ll}{
  \link[=tCorpus$data]{$data()} \tab Get a copy of the token data. Cannot be used to modify data (see Modify data section) \cr
  \link[=tCorpus$meta]{$meta()} \tab Get a copy of the document meta data \cr
  \link[=tCorpus$dtm]{$dtm()} \tab Create a document term matrix \cr
  \link[=tCorpus$feature_index]{$feature_index()} \tab Create or load a feature index. Used for creating functions/methods that require fast lookup of features. \cr
  \link[=tCorpus$context]{$context()} \tab Get a context vector: documents or globally unique sentences)
}
\strong{Modify}

the tCorpus data objects cannot be accessed directly, in order to protect the structure that many methods assume. The following methods can be used to modify the data while keeping this structure intact.
\tabular{ll}{
  \link[=tCorpus$transform]{$transform()} \tab Modify the token data using the \link{transform.data.table} function. \cr
  \link[=tCorpus$within]{$within()} \tab Modify the token data using the \link{within.data.table} function. \cr
  \link[=tCorpus$set_column]{$set_column()} \tab Modify the token data by setting the values of one (existing or new) column. \cr
  \link[=tCorpus$transform]{$transform_meta()} \tab The transform method for the document meta data \cr
  \link[=tCorpus$within]{$within_meta()} \tab The within method for the document meta data \cr
  \link[=tCorpus$set_column]{$set_meta_column()} \tab The set_column method for the document meta data
}
\strong{Subset / select / filter}
\tabular{ll}{
  \link[=tCorpus$subset]{$subset()} \tab Modify the token and/or meta data using the \link{subset.tCorpus} function. A subset expression can be specified for both the token data (subset) and the document meta data (subset_meta). \cr
  \link[=tCorpus$subset_query]{$subset_query()} \tab to be added
}
\strong{Check and clean}
\tabular{ll}{
  \link[=tCorpus$feature_index]{$reset_feature_index()} \tab Reset the feature index. (free up memory) \cr
  $set_keys() \tab Check and (re)set the data.table keys. Should not be necessary, but pointed out here till certain. \cr
  $droplevels() \tab Drop the unused factor levels in the token and meta data.
}
}

