% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_overview.r
\name{tCorpus_docsim}
\alias{tCorpus_docsim}
\title{Document similarity}
\description{
\link[=tCorpus]{(back to overview)}
}
\details{
\strong{Compare documents, and perform similarity based deduplication}
\tabular{ll}{
  \link[=compare_documents]{compare_documents()} \tab Compare documents \cr
  \link[=tCorpus$deduplicate]{$deduplicate()} \tab Remove duplicate documents
}
}
