% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_overview.r
\name{tCorpus_features}
\alias{tCorpus_features}
\title{Preprocessing, subsetting and analyzing features}
\description{
\link[=tCorpus]{(back to overview)}
}
\details{
\strong{Pre-process features}
\tabular{ll}{
  \link[=tCorpus$preprocess]{$preprocess()} \tab Create or modify a feature by preprocessing an existing feature \cr
  \link[=tCorpus$feature_subset]{$feature_subset()} \tab Similar to using subset, but instead of deleting rows it only sets rows for a specified feature to NA.
}
\strong{Inspect features}
\tabular{ll}{
  \link[=feature_stats]{feature_stats()} \tab Create a data.frame with feature statistics \cr
  \link[=top_features]{top_features()} \tab Show top features, optionally grouped by a given factor
}
}
