% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_overview.r
\name{tCorpus_querying}
\alias{tCorpus_querying}
\title{Use Boolean queries to analyze the tCorpus}
\description{
\link[=tCorpus]{(back to overview)}
}
\details{
\strong{Feature-level queries}
\tabular{ll}{
  \link[=tCorpus$search_features]{$search_features)} \tab Search for features based on keywords and conditions \cr
  \link[=tCorpus$search_recode]{$search_recode()} \tab Use the search_features query syntax to recode features \cr
  \link[=tCorpus$kwic]{$kwic()} \tab Get keyword-in-context (kwic) strings
}
\strong{Context-level queries}
\tabular{ll}{
  \link[=tCorpus$search_contexts]{$search_contexts()} \tab Search for documents or sentences using Lucene-like queries \cr
  \link[=tCorpus$subset_query]{$subset_query()} \tab use the search_contexts query syntax to subset the tCorpus
}
}

