% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_overview.r
\name{tCorpus_resources}
\alias{tCorpus_resources}
\title{Language resources}
\description{
\link[=tCorpus]{(back to overview)}
}
\details{
\strong{Named entity recognition}
\tabular{ll}{
  \link[=tCorpus$jrc_names]{$jrc_names()} \tab Multilingual named entity recognition (in news items) \cr
}
}
