% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_overview.r
\name{tCorpus_semnet}
\alias{tCorpus_semnet}
\title{Feature co-occurrence based semantic network analysis}
\description{
\link[=tCorpus]{(back to overview)}
}
\details{
\strong{Create networks}
\tabular{ll}{
  \link[=semnet]{semnet)} \tab Feature co-occurrence within contexts (documents, sentences) \cr
  \link[=semnet_window]{semnet_window()} \tab Feature co-occurrence within a specified token distance
}
\strong{Support functions for analyzing and visualizing the semantic network}
\tabular{ll}{
  \link[=ego_semnet]{ego_semnet()} \tab Create an ego network from an Igraph network \cr
  \link[=plot_semnet]{plot_semnet()} \tab Convenience function for visualizing an Igraph network, specialized for semantic networks
}
}
