% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_overview.r
\name{tCorpus_topmod}
\alias{tCorpus_topmod}
\title{Topic modeling}
\description{
\link[=tCorpus]{(back to overview)}
}
\details{
\strong{Train a topic model}
\tabular{ll}{
  \link[=tCorpus$lda_fit]{$lda_fit()} \tab Latent Dirichlet Allocation
}
}
