% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_information.r
\name{term_statistics}
\alias{term_statistics}
\title{Compute some useful corpus statistics for a dtm}
\usage{
term_statistics(tc, feature, context_level = c("document", "sentence"))
}
\arguments{
\item{dtm}{a document term matrix}
}
\value{
A data frame with rows corresponding to the terms in dtm and the statistics in the columns
}
\description{
Compute a number of useful statistics for filtering words: term frequency, idf, etc.
}

