% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{tokenGrepl}
\alias{tokenGrepl}
\title{A wrapper for grepl that takes multiple patterns. For efficiency, grepl is performed on unique words, and results for individual values are matched. Uses batches to go easy on memory (though sacrificing a bit of speed)}
\usage{
tokenGrepl(patterns, x, ignore.case = T, perl = F, batchsize = 25,
  useBytes = T)
}
\arguments{
\item{perl}{}
}
\value{
a logical vector
}
\description{
A wrapper for grepl that takes multiple patterns. For efficiency, grepl is performed on unique words, and results for individual values are matched. Uses batches to go easy on memory (though sacrificing a bit of speed)
}

