% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.r
\name{topics_per_document}
\alias{topics_per_document}
\title{Get the topics per document}
\usage{
topics_per_document(m, as.wordassignments = F)
}
\arguments{
\item{m}{A LDA model from the topicmodels package}

\item{as.wordassignments}{If True, return the topic occurence per document as the number of words assigned to the topic (instead of the posterior distribution)}
}
\value{
A data frame with rows corresponding to the terms in dtm and the statistics in the columns
}
\description{
Return a data frame containing article ids and topic occurence per document. Topic occurence can either be based on the posterior distrition (default) or based on wordassignments (as.wordassignments = T).
}

