% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpipe_tqueries.r
\name{udpipe_clause_tqueries}
\alias{udpipe_clause_tqueries}
\title{Get a list of tqueries for extracting who did what}
\usage{
udpipe_clause_tqueries(verbs = NULL, exclude_verbs = verb_lemma("quote"))
}
\arguments{
\item{verbs}{A character vector for specific verbs to use. By default uses all verbs (except for those specified in exclude_verbs)}

\item{exclude_verbs}{A character vector for specific verbs NOT to use. By default uses the verbs that indicate speech (that are used for extracting who said what, in \code{\link{udpipe_quote_tqueries}})}
}
\description{
An off-the-shelf list of tqueries for extracting subject-verb clauses. Designed for working with a tCorpus created with \code{\link{udpipe_tcorpus}}.
}
\examples{
udpipe_clause_tqueries()
}
