% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsyntax_tqueries.r
\name{udpipe_quote_tqueries}
\alias{udpipe_quote_tqueries}
\title{Get a list of tqueries for extracting quotes}
\usage{
udpipe_quote_tqueries(say_verbs = verb_lemma("quote"))
}
\arguments{
\item{say_verbs}{A character vector of verb lemma that indicate speech (e.g., say, state). A default list is included in verb_lemma('quote'), but certain lemma might be more accurate/appropriate depending on the corpus.}
}
\description{
An off-the-shelf list of tqueries for extracting quotes. Designed for working with a tCorpus created with \code{\link{udpipe_tcorpus}}.
}
\examples{
udpipe_quote_queries()
}
