% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_functions.r
\name{wordWindowOccurence}
\alias{wordWindowOccurence}
\title{Gives the window in which a term occured in a matrix.}
\usage{
wordWindowOccurence(tokenlist, window.size = 10, direction = "<>",
  distance.as.value = F, feature.col = tokenlist@feature.col, filter = T)
}
\arguments{
\item{window.size}{The distance within which words should occur from each other to be counted as a co-occurence.}

\item{direction}{a string indicating whether only the left ('<') or right ('>') side of the window, or both ('<>'), should be used.}
}
\value{
A list with two matrices. position.mat gives the specific position of a term, and window.mat gives the window in which each word occured. The rows represent the position of a term, and matches the input of this function (position, term and context). The columns represents terms.
}
\description{
This function returns the occurence of words (position.matrix) and the window of occurence (window.matrix). This format enables the co-occurence of words within sliding windows (i.e. word distance) to be calculated by multiplying position.matrix with window.matrix.
}

