% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenvis.r
\name{wrap_html}
\alias{wrap_html}
\title{Wrap a call that 'cats' html fragments and save it as a html file with header and footer}
\usage{
wrap_html(wrapped, header_info = NULL, template = "simple", output = NULL,
  width = 500)
}
\description{
Since R evaluates arguments lazily, call with the actual wrapped call, i.e. wrap_html(render_html(...))
}

