% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_results.r
\name{XnotY}
\alias{XnotY}
\title{Title}
\usage{
XnotY(hits.x, hits.y, doc.col = getOption("doc.col", "doc_id"),
  position.col = getOption("position.col", "position"))
}
\arguments{
\item{...}{}
}
\description{
Title
}

