% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocations.r
\name{add_collocation_label}
\alias{add_collocation_label}
\title{Choose and add collocation strings based on collocation categories}
\usage{
add_collocation_label(tc, colloc_id, feature = "word",
  new_feature = sprintf("\%s_l", colloc_id), pref_subset = NULL)
}
\arguments{
\item{tc}{a tcorpus object}

\item{colloc_id}{the data column containing the unique id for collocation words}
}
\description{
Given a collocation category (e.g., named entity ids), this function finds the most frequently occuring string in this category and adds it as a label for the category
}

