% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.r
\name{calc_chi2}
\alias{calc_chi2}
\title{Compute the chi^2 statistic for a 2x2 crosstab containing the values
[a, b]
[c, d]}
\usage{
calc_chi2(a, b, c, d, correct = T, cochrans_criteria = F)
}
\description{
Compute the chi^2 statistic for a 2x2 crosstab containing the values
[a, b]
[c, d]
}

