% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semnet.r
\name{cooccurrence_window}
\alias{cooccurrence_window}
\title{A sliding window approach to calculate the co-occurence of words}
\usage{
cooccurrence_window(tc, feature, context_level = c("document", "sentence"),
  window.size = 10, direction = "<>")
}
\arguments{
\item{window.size}{The distance within which words should occur from each other to be counted as a co-occurence.}

\item{direction}{a string indicating whether only the left ('<') or right ('>') side of the window, or both ('<>'), should be used.}
}
\value{
An edgelist (data.frame) with columns x, y and weight, in which weight represents the number of times y occured within a [window.size] word distance from x.
}
\description{
A sliding window approach to calculate the co-occurence of words
}

