% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcorpus.r
\name{create_tcorpus}
\alias{create_tcorpus}
\alias{create_tcorpus.character}
\alias{create_tcorpus.data.frame}
\alias{create_tcorpus.factor}
\title{Create a tCorpus}
\usage{
create_tcorpus(x, ...)

\method{create_tcorpus}{character}(x, doc_id = 1:length(x), meta = NULL,
  split_sentences = F, max_sentences = NULL, max_words = NULL,
  verbose = F)

\method{create_tcorpus}{factor}(x, doc_id = 1:length(x), meta = NULL,
  split_sentences = F, max_sentences = NULL, max_words = NULL,
  verbose = F)

\method{create_tcorpus}{data.frame}(d, text_columns = "text",
  doc_column = NULL, split_sentences = F, max_sentences = NULL,
  max_words = NULL)
}
\arguments{
\item{meta}{A data.frame with document meta information (e.g., date, source). The rows of the data.frame need to match the values of x}

\item{split_sentences}{Logical. If TRUE, the sentence number of tokens is also computed. This is done using the sentence tokenization of quanteda::tokenize, which is noted by the authors to be good (especially for english) but not perfect.}

\item{max_sentences}{An integer. Limits the number of sentences per document to the specified number. If set when split_sentences == FALSE, split_sentences will be set to TRUE.}

\item{max_words}{An integer. Limits the number of words per document to the specified number}
}
\description{
Create a tCorpus
}

