% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered_tcorpus.r
\name{delete_shattered_tc}
\alias{delete_shattered_tc}
\title{Delete a shattered_tCorpus and all files it refers to}
\usage{
delete_shattered_tc(stc)
}
\arguments{
\item{stc}{A shattered_tCorpus object}
}
\description{
Delete a shattered_tCorpus and all files it refers to
}

