% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.r
\name{download_resource}
\alias{download_resource}
\title{Download language resources}
\usage{
download_resource(resource = c("jrc_names"),
  local_path = getOption("tcorpus_resources", NULL), force_download = F)
}
\arguments{
\item{resource}{A character string indicating which resource to download}

\item{local_path}{The path (without the filename) where the resource is stored. If NULL (default), the file will be stored in the location of the tcorpus package.}

\item{force_download}{If TRUE, the resource will be downloaded even if it is already stored locally. Only usefull if resources have been updated.}
}
\description{
Currently available resources:
\itemize{
   \item{JRC-Names: "JRC-Names is a highly multilingual named entity resource for person and organisation names. [...] JRC-Names is a by-product of the analysis of about 220,000 news reports per day by the Europe Media Monitor (EMM) family of applications." (https://ec.europa.eu/jrc/en/language-technologies/jrc-names)}
}
}
\details{
The resource file is downloaded and stored on disk.
By default, the file will be saved to the location of the tcorpus package (for which you need write access).
Alternatively, the file can be saved as the specified local_path.
This can also be specified globally by using the set_resources_path function.
}
