% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_corpora.r
\name{dtm_compare}
\alias{dtm_compare}
\title{Compare two document term matrices}
\usage{
dtm_compare(dtm.x, dtm.y = NULL, smooth = 1, min_over = NULL,
  min_chi2 = NULL, select.rows = NULL, yates_cor = c("auto", "yes", "no"),
  x_is_subset = F)
}
\arguments{
\item{dtm.x}{the main document-term matrix}

\item{dtm.y}{the 'reference' document-term matrix}

\item{smooth}{the smoothing parameter for Laplace smoothing.}
}
\value{
A data frame with rows corresponding to the terms in dtm and the statistics in the columns
}
\description{
Compare two document term matrices
}

