% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.r
\name{feature_from_resource}
\alias{feature_from_resource}
\title{Use a language resource to create a feature}
\usage{
feature_from_resource(tc, resource = "jrc_names", new_feature = resource,
  feature = "word", resource_path = NULL, collocation_labels = T,
  verbose = T, batchsize = "default")
}
\arguments{
\item{tc}{a tCorpus object}

\item{resource}{A character string indicating which resource to use.}

\item{new_feature}{The name of the new feature. Default is the name of the resource}

\item{feature}{The feature to be used as input. Note that most resources assume that the input consists of unprocessed words}

\item{resource_path}{The path (without the filename) where the resource is stored. See ?download_resource for more information.}

\item{collocation_labels}{if True, then for resources that create an id for subsequent words (e.g., named entities), labels are added (in a separate column) based on the most frequent collocation combinations in 'your' data. Note that this means that the labels can be different if you run the same analysis on a different corpus; this is why the id is always kept.}

\item{force_download}{If TRUE, the resource will be downloaded even if it is stored locally.}
}
\description{
Currently available resources:
\itemize{
   \item{JRC-Names: "JRC-Names is a highly multilingual named entity resource for person and organisation names. [...] JRC-Names is a by-product of the analysis of about 220,000 news reports per day by the Europe Media Monitor (EMM) family of applications." (https://ec.europa.eu/jrc/en/language-technologies/jrc-names)}
}
}
\details{
The resource needs to be downloaded first. For this you can use the download_resource function, which will (by default) download the resource into the tcorpus package folder.
}

