% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenvis.r
\name{html_tagged_text}
\alias{html_tagged_text}
\title{Render text as HTML}
\usage{
html_tagged_text(d, doc_col = "doc_id", text_col = "word", meta = NULL,
  highlight = NULL, scale = NULL, group = NULL,
  highlight_col = "yellow", scale_col = c("red", "blue"), ...)
}
\arguments{
\item{d}{a data.frame with tokens}

\item{doc_col}{The column containing the document id}

\item{text_col}{The column that is used to create the output text}

\item{meta}{A data.frame with a doc_id columns (same name as doc_col). Meta columns are presented in the HTML as a table.}

\item{highlight}{A numeric or logical vector. Words that are TRUE, not NA and not zero will be highlighted. If highlight is a numeric vector, values determine the alpha of the highlight color (must be between 0 and 1).}

\item{scale}{Like highlight, but uses a color scale. Must be a numeric vector with values ranging from -1 to 1.}

\item{group}{A vector of length tc$n. Unique values will be used to assign words to categories. NA values can be used to leave words unassigned.}

\item{highlight_col}{a character vector of lenght 1 with a color name. Determines the color used in highlight}

\item{scale_col}{a character vector of length 2 with color names, giving the left and right color of the scale}

\item{...}{additional arguments to templates}
}
\description{
Render text as HTML
}

