% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.r
\name{preprocess_feature}
\alias{preprocess_feature}
\title{Preprocess feature}
\usage{
preprocess_feature(tc, column, new_column, language = "english",
  use_stemming = F, lowercase = T, ngrams = 1,
  ngram_context = c("document", "sentence"), remove_accented = F)
}
\arguments{
\item{ngram_context}{}
}
\description{
Preprocess feature
}

