% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_results.r
\name{print.contextHits}
\alias{print.contextHits}
\title{S3 print for contextHits class}
\usage{
\method{print}{contextHits}(x, ...)
}
\arguments{
\item{x}{a contextHits object, as returned by \link{search_contexts}}

\item{...}{not used}
}
\description{
S3 print for contextHits class
}
\examples{
text = c('A B C', 'D E F. G H I', 'A D', 'GGG')
tc = create_tcorpus(text, doc_id = c('a','b','c','d'), split_sentences = TRUE)
hits = search_contexts(tc, c('query label# A AND B', 'second query# (A AND Q) OR ("D E") OR I'))

hits
}
