% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shatter_tcorpus.r
\name{redistribute_shards}
\alias{redistribute_shards}
\title{Redistribute shards in a shattered tCorpus}
\usage{
redistribute_shards(stc, tokens_per_shard = 1e+05)
}
\arguments{
\item{stc}{A shattered_tCorpus object}

\item{tokens_per_shard}{The number of tokens per shard. This is an approximation, since a tCorpus should not be broken within documents. Also, given the number of shards that is required, tokens will be evently distributed (for instance, if tokens_per_shard is 1,000,000, and there are 1,500,000 tokens, you'll get 2 tokens of 750,000 shards)}
}
\description{
For good performance it's generally best to keep shards of moderate size. Not too small to prevent tedious merging and looping, and not too big to prevent memory problems. This function redistributes the shards (within categories) to approximate the specified tokens_per_shard
}
