% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semnet_plotting.r
\name{setNetworkAttributes}
\alias{setNetworkAttributes}
\title{Set some default network attributes for pretty plotting}
\usage{
setNetworkAttributes(g, size_attribute = "freq", color_attribute = NA,
  redo_layout = F)
}
\arguments{
\item{g}{A graph in the Igraph format.}

\item{size_attribute}{the name of the vertex attribute to be used to set the size of nodes}

\item{if}{TRUE, isolates are placed next to the network}
}
\value{
a network in the Igraph format
}
\description{
The purpose of this function is to create some default network attribute options to plot networks in a nice and insightfull way.
}

