% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_tcorpus.r
\name{subset.tCorpus}
\alias{subset.tCorpus}
\title{Subset a tCorpus}
\usage{
\method{subset}{tCorpus}(tc, subset = NULL, subset_meta = NULL,
  drop_levels = F, window = NULL)
}
\arguments{
\item{tc}{tCorpus object}

\item{subset}{logical expression indicating elements or rows to keep in the tokens data.}

\item{window}{If not NULL, an integer specifiying the window to be used to return the subset. For instance, if the subset contains word 10 in a document and window is 5, the subset will contain word 5 to 15. Naturally, this does not apply to subset_meta.}
}
\description{
Returns the subset of a tCorpus. The selection can be made separately (and simultaneously) for the token data (using subset) and the meta data (using subset_meta).
}

