% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_results.r
\name{summary.contextHits}
\alias{summary.contextHits}
\title{S3 summary for contextHits class}
\usage{
\method{summary}{contextHits}(object, ...)
}
\arguments{
\item{object}{a contextHits object, as returned by \link{search_contexts}}

\item{...}{not used}
}
\description{
S3 summary for contextHits class
}
\examples{
text = c('A B C', 'D E F. G H I', 'A D', 'GGG')
tc = create_tcorpus(text, doc_id = c('a','b','c','d'), split_sentences = TRUE)
hits = search_contexts(tc, c('query label# A AND B', 'second query# (A AND Q) OR ("D E") OR I'))

summary(hits)
}
