% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_results.r
\name{summary.featureHits}
\alias{summary.featureHits}
\title{S3 summary for featureHits class}
\usage{
\method{summary}{featureHits}(object, ...)
}
\arguments{
\item{object}{a featureHits object, as returned by \link{search_features}}

\item{...}{not used}
}
\description{
S3 summary for featureHits class
}
\examples{
text = c('A B C', 'D E F. G H I', 'A D', 'GGG')
tc = create_tcorpus(text, doc_id = c('a','b','c','d'), split_sentences = TRUE)
hits = search_features(tc, c('query label# A AND B', 'second query# (A AND Q) OR ("D E") OR I'))

summary(hits)
}
