% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$data}
\alias{data.tCorpus}
\alias{meta.tCorpus}
\alias{tCorpus$data}
\alias{tCorpus$meta}
\title{Access the data from a tCorpus}
\usage{
## R6 method for class tCorpus.
\code{tCorpus$data}
\code{tCorpus$meta}
}
\description{
Several limitations:
- subsetting or appending the data is not possible. (use the subset function and merge_tcorpua)
- data.table operations that change the data.table by reference. Note that the set_column and set_meta_column functions do use assignment by reference.
}

