% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$data}
\alias{data.tCorpus}
\alias{meta.tCorpus}
\alias{tCorpus$data}
\alias{tCorpus$meta}
\title{Extract the data from a tCorpus}
\usage{
## R6 method for class tCorpus.
\code{tCorpus$data(columns=NULL, keep_df=F, as.df=F)}
\code{tCorpus$meta(columns=NULL, keep_df=F, as.df=F, per_token)}
}
\arguments{
\item{columns}{A character string indicating which columns to return. NULL means all columns.}

\item{keep_df}{Keep the data.table/data.frame if only one column is selected.}

\item{as.df}{Return a regular data.frame instead of a data.table}

\item{per_token}{Repeat rows in the document meta data so that it matches with the token data}
}
\description{
Extract the data from a tCorpus
}

