% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$data}
\alias{data.tCorpus}
\alias{meta.tCorpus}
\alias{tCorpus$data}
\alias{tCorpus$meta}
\title{Access specific column from the data or meta data.}
\description{
Alternative to using the $data and $meta field. This is more efficient if not all columns are selected, because $data copies the whole data.table, whereas get only copies the selected columns.
}
\section{Usage}{

## R6 method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{get}
\preformatted{meta}
}

