% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$data}
\alias{data.tCorpus}
\alias{meta.tCorpus}
\alias{tCorpus$data}
\alias{tCorpus$meta}
\title{Access the data from a tCorpus}
\usage{
## R6 method for class tCorpus.
}
\description{
The $data and $meta fields can be used to access and modify the token data and meta data.

Several limitations:
- subsetting or appending the data is not possible. (use the subset function and merge_tcorpua)
- data.table operations that change the data.table by reference. Note that the set and set_meta functions do use assignment by reference.
}

