% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$delete_columns}
\alias{tCorpus$delete_columns}
\alias{delete_columns.tCorpus}
\alias{tCorpus$delete_meta_columns}
\alias{delete_meta_columns.tCorpus}
\title{Delete column from the data and meta data}
\arguments{
\item{cnames}{the names of the columns to delete}
}
\description{
Delete column from the data and meta data
}
\section{Usage}{

## R6 method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{delete_columns(cnames)}
\preformatted{delete_meta_columns(cnames)}
}

