% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$delete_columns}
\alias{tCorpus$delete_columns}
\alias{tCorpus$delete_meta_columns}
\alias{delete_columns}
\alias{delete_meta_columns}
\title{Delete column from the data and meta data}
\arguments{
\item{cnames}{the names of the columns to delete}
}
\description{
\strong{Usage:}
}
\details{
## R6 method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{delete_columns(cnames)}
\preformatted{delete_meta_columns(cnames)}
}
\examples{
d = data.frame(text = c('Text one','Text two','Text three'),
               date = c('2010-01-01','2010-01-01','2012-01-01'))
tc = create_tcorpus(d)

tc$tokens
tc$delete_columns('token')
tc$tokens

tc$meta
tc$delete_meta_columns('date')
tc$meta
}
