% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$feature_stats}
\alias{feature_stats.tCorpus}
\alias{tCorpus$feature_stats}
\title{Feature statistics}
\usage{
## R6 method for class tCorpus
}
\arguments{
\item{feature}{The name of the feature}

\item{sent_freq}{If True, include sentence frequency (only if sentence information is available).}
}
\description{
Compute a number of useful statistics for features: term frequency, idf, etc.
}

