% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$get}
\alias{tCorpus$get}
\alias{get.tCorpus}
\alias{tCorpus$get_meta}
\alias{get_meta.tCorpus}
\title{Access the data from a tCorpus}
\arguments{
\item{a}{character vector with the names of the columns}

\item{keep_df}{if True, the output will be a data.table (or data.frame) even if it only contains 1 columns}

\item{as.df}{if True, the output will be a regular data.frame instead of a data.table}
}
\description{
Alternative to using the $data and $meta field. This is more efficient if not all columns are selected, because $data copies the whole data.table, whereas get only copies the selected columns.
}
\section{Usage}{

## R6 active method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{get(columns=NULL, keep_df=F, as.df=F)}
\preformatted{get_meta(columns=NULL, keep_df=F, as.df=F)}
}

