% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$get}
\alias{tCorpus$get}
\alias{get.tCorpus}
\alias{tCorpus$get_meta}
\alias{get_meta.tCorpus}
\title{Access the data from a tCorpus}
\arguments{
\item{a}{character vector with the names of the columns}

\item{keep_df}{if True, the output will be a data.table (or data.frame) even if it only contains 1 columns}

\item{as.df}{if True, the output will be a regular data.frame instead of a data.table}
}
\description{
The $data and $meta fields can be used to access and modify the token data and meta data.

Several limitations:
- subsetting or appending the data is not possible. (use the subset function and merge_tcorpua)
- data.table operations that change the data.table by reference are not possible. Note that the set and set_meta functions do use assignment by reference.
- the returned data.table is a copy of the data.table within the corpus. This copy is an actual copy (the copy-on-modify mechanic doesn't work here) which can eat up memory if. Accordingly, if you only want to get 1 or several columns, it is more efficient to use the get() and get_meta() methods, which only copy the selected columns.
}
\section{Usage}{

## R6 active method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{get(columns=NULL, keep_df=F, as.df=F)}
\preformatted{get_meta(columns=NULL, keep_df=F, as.df=F)}
}

