% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$search_recode}
\alias{search_recode.tCorpus}
\alias{tCorpus$search_recode}
\title{Recode features in a tCorpus based on a search string}
\usage{
## R6 method for class tCorpus
\code{tCorpus$search_recode(feature, new_value, keyword, condition=NA, condition_once=F, subset_tokens=NA, subset_meta=NA, clone=self$clone_on_change)}
}
\arguments{
\item{feature}{The feature in which to search}

\item{new_value}{the character string with which all features that are found are replaced}

\item{...}{See \link{tCorpus$search_features} for the query parameters}

\item{clone}{If TRUE, the method returns a new tCorpus object. This is the normal R way of doing things. Alternatively, the tCorpus can be used as a reference class object by setting clone to FALSE, or setting tCorpus$clone_on_change to FALSE to use this globally. Please consult the general documentation for tCorpus (?tCorpus) for a more detailed explanation.}
}
\description{
Search features (see \link{tCorpus$search_features}) and replace features with a new value
}

