% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_features.r
\name{tCorpus$search_recode}
\alias{tCorpus$search_recode}
\alias{search_recode}
\title{Recode features in a tCorpus based on a search string}
\arguments{
\item{feature}{The feature in which to search}

\item{new_value}{the character string with which all features that are found are replaced}

\item{query}{See \link{tCorpus$search_features} for the query parameters}

\item{...}{Additional search_features parameters. See \link{tCorpus$search_features}}
}
\description{
Search features (see \link{tCorpus$search_features}) and replace features with a new value

\strong{Usage:}

## R6 method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{
search_recode(feature, new_value, keyword, condition = NA, condition_once = F)
}
}
