% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$set_levels}
\alias{tCorpus$set_levels}
\alias{set_levels.tCorpus}
\alias{tCorpus$set_meta_levels}
\alias{set_meta_levels.tCorpus}
\title{Change levels of factor columns}
\arguments{
\item{column}{the name of the column}

\item{levels}{The new levels}
}
\description{
Change levels of factor columns
}
\section{Usage}{

## R6 method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{set_levels(column, levels)}
\preformatted{set_meta_levels(column, levels)}
}

