% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$subset}
\alias{subset.tCorpus}
\alias{tCorpus$subset}
\title{Subset a tCorpus}
\usage{
## R6 method for class tCorpus
\code{tCorpus$subset(subset=NULL, subset_meta=NULL, drop_levels=F, window=NULL)}
}
\arguments{
\item{subset}{logical expression indicating rows to keep in the tokens data.}

\item{subset_meta}{logical expression indicating rows to keep in the document meta data.}

\item{drop_levels}{if TRUE, drop all unused factor levels after subsetting}

\item{window}{If not NULL, an integer specifiying the window to be used to return the subset. For instance, if the subset contains word 10 in a document and window is 5, the subset will contain word 5 to 15. Naturally, this does not apply to subset_meta.}

\item{clone}{If TRUE, the method returns a new tCorpus object. This is the normal R way of doing things. Alternatively, the tCorpus can be used as a reference class object by setting clone to FALSE, or setting tCorpus$clone_on_change to FALSE to use this globally. Please consult the general documentation for tCorpus (?tCorpus) for a more detailed explanation.}
}
\description{
Returns the subset of a tCorpus. The selection can be made separately (and simultaneously) for the token data (using subset) and the meta data (using subset_meta). The subset arguments work according to the \link{subset.data.table} function.

## add documentation for freq(), docfreq() etc!!!

A minor difference with the subset.data.table function is that here it is also allowed to select a subset based on row indices, by only providing a numerical vector as argument.
}

