% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$subset}
\alias{subset.tCorpus}
\alias{tCorpus$subset}
\title{Subset a tCorpus}
\usage{
## R6 method for class tCorpus
}
\arguments{
\item{subset}{logical expression indicating rows to keep in the tokens data.}

\item{subset_meta}{logical expression indicating rows to keep in the document meta data.}

\item{drop_levels}{if TRUE, drop all unused factor levels after subsetting}

\item{window}{If not NULL, an integer specifiying the window to be used to return the subset. For instance, if the subset contains word 10 in a document and window is 5, the subset will contain word 5 to 15. Naturally, this does not apply to subset_meta.}

\item{copy}{If TRUE, the method returns a new tCorpus object. This is the normal R way of doing things. Alternatively, the tCorpus can be used as a reference class object by setting copy to FALSE, or setting tCorpus$copy_on_modify to FALSE to use this globally. Please consult the general documentation for tCorpus (?tCorpus) for a more detailed explanation.}
}
\description{
Returns the subset of a tCorpus. The selection can be made separately (and simultaneously) for the token data (using subset) and the meta data (using subset_meta). The subset arguments work according to the \link{subset.data.table} function.

Subset can also be used to select rows based on word/feature frequences. This is a common step in corpus analysis, where it often makes sense to ignore very rare and/or very frequent words.
To do so, there are several special functions that can be used within a subset call.
The freq_filter() and docfreq_filter() can be used to filter terms based on term frequency and document frequency, respectively.
The first argument to these functions is the name of the feature, e.g., freq_filter(word).
To filter this feature you can specify: min (minimum frequency), max (maximum frequency), top (n most frequent words) and bottom (n least frequent words).
For example, tc$freq_filter(word, min=10) filters out all rows in which the specified feature occures at least 10 times in the entire corpus.
tc$freq_filter(word, top=100) keeps only the rolws in which the feature is in the top 100 most frequent features.
tc$docfreq_filter(word, max = 0.9 * tc$n) deletes all rows with features that occured in more than 90% of all documents.

Note that you can also use the \link{tCorpus$feature_subset} method if you want to filter out low/high frequency words, but do not want to delete the rows in the tCorpus.
}

