% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$subset_query}
\alias{tCorpus$subset_query}
\alias{subset_query.tCorpus}
\title{Subset tCorpus token data using a query}
\arguments{
\item{copy}{If TRUE, the method returns a new tCorpus object. This is the normal R way of doing things. Alternatively, the tCorpus can be used as a reference class object by setting copy to FALSE, or setting tCorpus$always_copy to FALSE to use this globally. Please consult the general documentation for tCorpus (?tCorpus) for a more detailed explanation.}
}
\description{
A convenience function that searches for contexts (documents, sentences), and uses the results to \link[=tCorpus$search_contexts]{subset} the tCorpus token data.

See the documentation for \link[=tCorpus$search_contexts]{subset} for an explanation of the query language.
}
\section{Usage}{

## R6 method for class tCorpus. Use as tc$method (where tc is a tCorpus object).

\preformatted{subset_query(query, feature = 'token', context_level = c('document','sentence'), copy = self$always_copy)}
}

