% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_methods.r
\name{tCorpus$subset_query}
\alias{subset_query.tCorpus}
\alias{tCorpus$subset_query}
\title{Subset tCorpus token data using a query}
\usage{
## R6 method for class tCorpus
}
\arguments{
\item{copy}{If TRUE, the method returns a new tCorpus object. This is the normal R way of doing things. Alternatively, the tCorpus can be used as a reference class object by setting copy to FALSE, or setting tCorpus$copy_on_modify to FALSE to use this globally. Please consult the general documentation for tCorpus (?tCorpus) for a more detailed explanation.}
}
\description{
A convenience function that searches for contexts (documents, sentences), and uses the results to \link[=tCorpus$search_contexts]{subset} the tCorpus token data.

See the documentation for \link[=tCorpus$search_contexts]{subset} for an explanation of the query language.
}

