% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsyntax_tqueries.r
\name{tCorpus$udpipe_clauses}
\alias{tCorpus$udpipe_clauses}
\alias{udpipe_clauses}
\title{Add columns indicating who did what}
\arguments{
\item{column}{The name of the column in $tokens to store the results.}

\item{tqueries}{A list of tQueries. By default uses the off-the-shelf tqueries in \code{\link{udpipe_clause_tqueries}}}
}
\value{
a tCorpus
}
\description{
An off-the-shelf application of rsyntax for extracting subject-verb clauses. Designed for working with a tCorpus created with \code{\link{udpipe_tcorpus}}.
}
\examples{
\donttest{
tc = tc_sotu_udpipe$copy()
tc$udpipe_clauses()
tc_plot_tree(tc, token, lemma, POS, annotation='clause')
tc_syntax_reader(tc, annotation='clause', value='subject')
}
}
