% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpipe_coref.r
\name{tCorpus$udpipe_coref}
\alias{tCorpus$udpipe_coref}
\alias{udpipe_coref}
\title{Perform rule based coreference resolution}
\arguments{
\item{keep_feats}{If TRUE, keep the feature columns created for the coref resolution}
}
\value{
a tCorpus
}
\description{
This is an experimental implementation of coreference resolution, aimed at coreferences of nouns and pronouns.
}
\examples{
\donttest{
tc = tc_sotu_udpipe$copy()
tc$annotate_clauses()
tc_plot_tree(tc, token, lemma, POS, annotation='clause')
browse_texts(tc, rsyntax='clause', value='subject')
}
}
