% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsyntax_tqueries.r
\name{tCorpus$udpipe_quotes}
\alias{tCorpus$udpipe_quotes}
\alias{udpipe_quotes}
\title{Add columns indicating who said what}
\arguments{
\item{tqueries}{A list of tQueries. By default uses the off-the-shelf tqueries in \code{\link{udpipe_quote_tqueries}}}

\item{span_quotes}{If True, also look for quotes indicated with quotation marks, that can span multiple sentences. The source will be the most recent source in the previous 2 sentences, or if there is none, simple tqueries are used to find the most recent proper names that said/did something.}

\item{say_verbs}{If span_quotes is used, say_verbs can be required to find sources. say_verbs should be a character vector of verb lemma that indicate speech (e.g., say, state). A default list is included in verb_lemma('quote'), but certain lemma might be more accurate/appropriate depending on the corpus.}
}
\value{
the columns 'quote', 'quote_id', and 'quote_verbatim' are added to tokens
}
\description{
An off-the-shelf application of rsyntax for extracting quotes. Designed for working with a tCorpus created with \code{\link{udpipe_tcorpus}}.
}
\examples{
\donttest{
txt = 'Bob said that he likes Mary. John did not like that: 
       "how dare he!". "It is I, John, who likes Mary!!"'
tc = udpipe_tcorpus(txt, model = 'english-ewt')
tc$udpipe_quotes()
tc_plot_tree(tc, token, lemma, POS, annotation='quote')
browse_texts(tc, rsyntax='quote', value='source')
}
}
