% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_overview.r
\name{tCorpus_compare}
\alias{tCorpus_compare}
\title{Corpus comparison}
\description{
\link[=tCorpus]{(back to overview)}
}
\details{
\strong{Compare vocabulary of two corpora}
\tabular{ll}{
  \link[=compare_corpus]{compare_corpus()} \tab Compare vocabulary of one tCorpus to another \cr
  \link[=compare_subset]{compare_subset()} \tab Compare subset of a tCorpus to the rest of the tCorpus
}
}
