% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_overview.r
\name{tCorpus_querying}
\alias{tCorpus_querying}
\title{Use Boolean queries to analyze the tCorpus}
\description{
\link[=tCorpus]{(back to overview)}
}
\details{
\strong{Feature-level queries}
\tabular{ll}{
  \link[=search_features]{search_features())} \tab Search for features based on keywords and conditions \cr
  \link[=tCorpus$code_features]{$code_features())} \tab  Add a column to the token data based on feature search results \cr
  \link[=tCorpus$search_recode]{$search_recode()} \tab Use the search_features query syntax to recode features \cr
  \link[=feature_associations]{feature_associations()} \tab Given a query, get words that often co-occur nearby \cr
  \link[=get_kwic]{kwic()} \tab Get keyword-in-context (kwic) strings \cr
  \link[=browse_hits]{browse_hits()} \tab Create full-text browsers with highlighted search hits
}
\strong{Context-level queries}
\tabular{ll}{
  \link[=search_contexts]{search_contexts()} \tab Search for documents or sentences using Lucene-like queries \cr
  \link[=tCorpus$subset_query]{$subset_query()} \tab use the search_contexts query syntax to subset the tCorpus
}
}
