% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcorpus.r
\name{tokens_to_tcorpus}
\alias{tokens_to_tcorpus}
\title{Create a tcorpus based on tokens (i.e. preprocessed texts)}
\usage{
tokens_to_tcorpus(tokens, doc_col = "doc_id", word_i_col = NULL,
  sent_i_col = NULL, doc_meta = NULL, feature_cols = NULL)
}
\arguments{
\item{feature_cols}{}
}
\description{
Create a tcorpus based on tokens (i.e. preprocessed texts)
}

