% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semnet_plotting.r
\name{top_cluster_features}
\alias{top_cluster_features}
\title{Get top features from semnet clusters}
\usage{
top_cluster_features(g, cluster_attr, measure = c("degree", "freq"),
  top_n = 5)
}
\description{
Get top features from semnet clusters
}

